Module tantivy::aggregation::agg_result

Contains the final aggregation tree. This tree can be converted via the into() method from IntermediateAggregationResults. This conversion computes the final result. For example: The intermediate result contains intermediate average results, which is the sum and the number of values. The actual average is calculated on the step from intermediate to final aggregation result tree.

Structs

AggregationResults

The final aggegation result.

BucketEntry

This is the default entry for a bucket, which contains a key, count, and optionally sub_aggregations.

RangeBucketEntry

This is the range entry for a bucket, which contains a key, count, and optionally sub_aggregations.

Enums

AggregationResult

An aggregation is either a bucket or a metric.

BucketEntries

This is the wrapper of buckets entries, which can be vector or hashmap depending on if it’s keyed or not.

BucketResult

BucketEntry holds bucket aggregation result types.

MetricResult

MetricResult