Struct tantivy::aggregation::bucket::HistogramAggregation

pub struct HistogramAggregation {
    pub field: String,
    pub interval: f64,
    pub offset: Option<f64>,
    pub min_doc_count: Option<u64>,
    pub hard_bounds: Option<HistogramBounds>,
    pub extended_bounds: Option<HistogramBounds>,
    pub keyed: bool,
}

Histogram is a bucket aggregation, where buckets are created dynamically for given interval. Each document value is rounded down to its bucket.

E.g. if we have a price 18 and an interval of 5, the document will fall into the bucket with the key 15. The formula used for this is: ((val - offset) / interval).floor() * interval + offset

For this calculation all fastfield values are converted to f64.

Returned Buckets

By default buckets are returned between the min and max value of the documents, including empty buckets. Setting min_doc_count to != 0 will filter empty buckets.

The value range of the buckets can bet extended via extended_bounds or limit the range via hard_bounds.

Result

Result type is BucketResult with BucketEntry on the AggregationCollector.

Result type is IntermediateBucketResult with IntermediateHistogramBucketEntry on the DistributedAggregationCollector.

Limitations/Compatibility

JSON Format

{
    "prices": {
        "histogram": {
            "field": "price",
            "interval": 10
        }
    }
}

Response See BucketEntry

Fields
field: String

The field to aggregate on.

interval: f64

The interval to chunk your data range. Each bucket spans a value range of [0..interval). Must be a positive value.

offset: Option<f64>

Intervals implicitly defines an absolute grid of buckets [interval * k, interval * (k + 1)).

Offset makes it possible to shift this grid into [offset + interval * k, offset + interval * (k + 1)). Offset has to be in the range [0, interval).

As an example, if there are two documents with value 9 and 12 and interval 10.0, they would fall into the buckets with the key 0 and 10. With offset 5 and interval 10, they would both fall into the bucket with they key 5 and the range [5..15)

min_doc_count: Option<u64>

The minimum number of documents in a bucket to be returned. Defaults to 0.

hard_bounds: Option<HistogramBounds>

Limits the data range to [min, max] closed interval.

This can be used to filter values if they are not in the data range.

hard_bounds only limits the buckets, to force a range set both extended_bounds and hard_bounds to the same range.

Example

{
    "prices": {
       "histogram": {
           "field": "price",
           "interval": 10,
           "hard_bounds": {
               "min": 0,
               "max": 100
           }
       }
   }
}

extended_bounds: Option<HistogramBounds>

Can be set to extend your bounds. The range of the buckets is by default defined by the data range of the values of the documents. As the name suggests, this can only be used to extend the value range. If the bounds for min or max are not extending the range, the value has no effect on the returned buckets.

Cannot be set in conjunction with min_doc_count > 0, since the empty buckets from extended bounds would not be returned.

keyed: bool

Whether to return the buckets as a hash map

Implementations

impl HistogramAggregation

pub fn min_doc_count(&self) -> u64

Returns the minimum number of documents required for a bucket to be returned.