Module tantivy::aggregation::intermediate_agg_result

Contains the intermediate aggregation tree, that can be merged. Intermediate aggregation results can be used to merge results between segments or between indices.

Structs

IntermediateAggregationResults

Contains the intermediate aggregation result, which is optimized to be merged with other intermediate results.

IntermediateHistogramBucketEntry

This is the histogram entry for a bucket, which contains a key, count, and optionally sub_aggregations.

IntermediateRangeBucketEntry

This is the range entry for a bucket, which contains a key, count, and optionally sub_aggregations.

IntermediateRangeBucketResult

Range aggregation including error counts

IntermediateTermBucketEntry

This is the term entry for a bucket, which contains a count, and optionally sub_aggregations.

IntermediateTermBucketResult

Term aggregation including error counts

Enums

IntermediateAggregationResult

An aggregation is either a bucket or a metric.

IntermediateBucketResult

The intermediate bucket results. Internally they can be easily merged via the keys of the buckets.

IntermediateMetricResult

Holds the intermediate data for metric results