Module tantivy::analyzer

Structs

AnalyzerManager
The analyzer manager serves as a store for all of the configured analyzers.

JapaneseTokenizer
Simple japanese tokenizer based on the tinysegmenter crate.

LowerCaser
Token filter that lowercase terms.

RawTokenizer

RemoveLongFilter
RemoveLongFilter removes tokens that are longer than a given number of bytes (in UTF-8 representation).

SimpleTokenizer

Stemmer

Token
Token

TokenStreamChain

Traits

Analyzer

BoxedAnalyzer

TokenFilterFactory

TokenStream

Functions

box_analyzer