Struct tantivy::collector::MultiCollector

pub struct MultiCollector<'a> { /* private fields */ }

Multicollector makes it possible to collect on more than one collector. It should only be used for use cases where the Collector types is unknown at compile time.

If the type of the collectors is known, you can just group yours collectors in a tuple. See the Combining several collectors section of the collector documentation.

use tantivy::collector::{Count, TopDocs, MultiCollector};
use tantivy::query::QueryParser;
use tantivy::schema::{Schema, TEXT};
use tantivy::{doc, Index};

let mut schema_builder = Schema::builder();
let title = schema_builder.add_text_field("title", TEXT);
let schema = schema_builder.build();
let index = Index::create_in_ram(schema);
let mut index_writer = index.writer(3_000_000)?;
index_writer.add_document(doc!(title => "The Name of the Wind"))?;
index_writer.add_document(doc!(title => "The Diary of Muadib"))?;
index_writer.add_document(doc!(title => "A Dairy Cow"))?;
index_writer.add_document(doc!(title => "The Diary of a Young Girl"))?;
index_writer.commit()?;

let reader = index.reader()?;
let searcher = reader.searcher();

let mut collectors = MultiCollector::new();
let top_docs_handle = collectors.add_collector(TopDocs::with_limit(2));
let count_handle = collectors.add_collector(Count);
let query_parser = QueryParser::for_index(&index, vec![title]);
let query = query_parser.parse_query("diary").unwrap();
let mut multi_fruit = searcher.search(&query, &collectors).unwrap();

let count = count_handle.extract(&mut multi_fruit);
let top_docs = top_docs_handle.extract(&mut multi_fruit);

assert_eq!(count, 2);
assert_eq!(top_docs.len(), 2);

Implementations

impl<'a> MultiCollector<'a>

pub fn new() -> Self

Create a new MultiCollector

pub fn add_collector<'b: 'a, TCollector: Collector + 'b>(
&mut self,
collector: TCollector
) -> FruitHandle<TCollector::Fruit>

Add a new collector to our MultiCollector.