Trait tantivy::collector::ScoreTweaker

pub trait ScoreTweaker<TScore>: Sync {
    type Child: ScoreSegmentTweaker<TScore>;

    fn segment_tweaker(
        &self,
        segment_reader: &SegmentReader
    ) -> Result<Self::Child>;
}

ScoreTweaker makes it possible to tweak the score emitted by the scorer into another one.

The ScoreTweaker itself does not make much of the computation itself. Instead, it helps constructing Self::Child instances that will compute the score at a segment scale.

Required Associated Types

type Child: ScoreSegmentTweaker<TScore>

Type of the associated ScoreSegmentTweaker.

Required Methods

fn segment_tweaker(&self, segment_reader: &SegmentReader) -> Result<Self::Child>

Builds a child tweaker for a specific segment. The child scorer is associated with a specific segment.