Trait tantivy::collector::SegmentCollector

pub trait SegmentCollector: 'static {
    type Fruit: Fruit;

    fn collect(&mut self, doc: DocId, score: Score);
    fn harvest(self) -> Self::Fruit;
}

The SegmentCollector is the trait in charge of defining the collect operation at the scale of the segment.

.collect(doc, score) will be called for every documents matching the query.

Required Associated Types

type Fruit: Fruit

Fruit is the type for the result of our collection. e.g. usize for the Count collector.

Required Methods

fn collect(&mut self, doc: DocId, score: Score)

The query pushes the scored document to the collector via this method.

fn harvest(self) -> Self::Fruit

Extract the fruit of the collection from the SegmentCollector.