Enum tantivy::directory::ReadOnlySource

pub enum ReadOnlySource {
    Mmap(MmapReadOnly),
    Anonymous(SharedVecSlice),
}

Read object that represents files in tantivy.

These read objects are only in charge to deliver the data in the form of a constant read-only &[u8]. Whatever happens to the directory file, the data hold by this object should never be altered or destroyed.

Variants
Mmap(MmapReadOnly)

Mmap source of data

Anonymous(SharedVecSlice)

Wrapping a Vec<u8>

Methods

impl ReadOnlySource

pub fn empty() -> ReadOnlySource

Creates an empty ReadOnlySource

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn as_slice(&self) -> &[u8]

Returns the data underlying the ReadOnlySource object.

pub fn split(self, addr: usize) -> (ReadOnlySource, ReadOnlySource)

Splits into 2 ReadOnlySource, at the offset given as an argument.

pub fn slice(&self, from_offset: usize, to_offset: usize) -> ReadOnlySource

Creates a ReadOnlySource that is just a view over a slice of the data.

Keep in mind that any living slice extends the lifetime of the original ReadOnlySource,

For instance, if ReadOnlySource wraps 500MB worth of data in anonymous memory, and only a 1KB slice is remaining, the whole 500MBs are retained in memory.

pub fn slice_from(&self, from_offset: usize) -> ReadOnlySource

Like .slice(...) but enforcing only the from boundary.

Equivalent to .slice(from_offset, self.len())

pub fn slice_to(&self, to_offset: usize) -> ReadOnlySource

Like .slice(...) but enforcing only the to boundary.

Equivalent to .slice(0, to_offset)