Enum tantivy::directory::error::DeleteError

pub enum DeleteError {
    FileDoesNotExist(PathBuf),
    IOError(IOError),
}

Error that may occur when trying to delete a file

Variants
FileDoesNotExist(PathBuf)

The file does not exists.

IOError(IOError)

Any kind of IO error that happens when interacting with the underlying IO device.