Enum tantivy::directory::error::Incompatibility

pub enum Incompatibility {
    CompressionMismatch {
        library_compression_format: String,
        index_compression_format: String,
    },
    IndexMismatch {
        library_version: Version,
        index_version: Version,
    },
}

Type of index incompatibility between the library and the index found on disk Used to catch and provide a hint to solve this incompatibility issue

Variants

CompressionMismatch

Fields

library_compression_format: String

Compression algorithm used by the current version of tantivy

index_compression_format: String

Compression algorithm that was used to serialise the index

This library cannot decompress the index found on disk

IndexMismatch

Fields

library_version: Version

Version used by the library

index_version: Version

Version the index was built with

The index format found on disk isn’t supported by this version of the library