Enum tantivy::directory::error::OpenReadError

pub enum OpenReadError {
    FileDoesNotExist(PathBuf),
    IOError(IOError),
}

Error that may occur when accessing a file read

Variants
FileDoesNotExist(PathBuf)

The file does not exists.

IOError(IOError)

Any kind of IO error that happens when interacting with the underlying IO device.