Enum tantivy::directory::error::OpenWriteError

pub enum OpenWriteError {
    FileAlreadyExists(PathBuf),
    IOError(IOError),
}

Error that may occur when starting to write in a file

Variants
FileAlreadyExists(PathBuf)

Our directory is WORM, writing an existing file is forbidden. Checkout the Directory documentation.

IOError(IOError)

Any kind of IO error that happens when writing in the underlying IO device.