Struct tantivy::directory::Lock

pub struct Lock {
    pub filepath: PathBuf,
    pub is_blocking: bool,
}

A directory lock.

A lock is associated with a specific path.

The lock will be passed to Directory::acquire_lock.

Tantivy itself uses only two locks but client application can use the directory facility to define their own locks.

INDEX_WRITER_LOCK

META_LOCK

Check out these locks documentation for more information.

Fields
filepath: PathBuf

The lock needs to be associated with its own file path. Depending on the platform, the lock might rely on the creation and deletion of this filepath.

is_blocking: bool

is_blocking describes whether acquiring the lock is meant to be a blocking operation or a non-blocking.

Acquiring a blocking lock blocks until the lock is available.

Acquiring a non-blocking lock returns rapidly, either successfully or with an error signifying that someone is already holding the lock.