Struct tantivy::directory::ManagedDirectory

pub struct ManagedDirectory { /* fields omitted */ }

Wrapper of directories that keeps track of files created by Tantivy.

A managed directory is just a wrapper of a directory that keeps a (persisted) list of the files that have been created (and not deleted) by tantivy so far.

Thanks to this list, it implements a garbage_collect method that removes the files that were created by tantivy and are not useful anymore.

Methods

impl ManagedDirectory

fn new<Dir: Directory>(directory: Dir) -> Result<ManagedDirectory, Error>

Wraps a directory as managed directory.

fn garbage_collect<L: FnOnce() -> HashSet<PathBuf>>(
&mut self,
get_living_files: L
)

Garbage collect unused files.

Removes the files that were created by tantivy and are not used by any segment anymore.

living_files - List of files that are still used by the index.

This method does not panick nor returns errors. If a file cannot be deleted (for permission reasons for instance) an error is simply logged, and the file remains in the list of managed files.

fn protect_file_from_delete(&self, path: &Path) -> FileProtection

Protects a file from being garbage collected.

The method returns a FileProtection object. The file will not be garbage collected as long as the FileProtection object is kept alive.