Struct tantivy::directory::MmapDirectory

pub struct MmapDirectory { /* fields omitted */ }

Directory storing data in files, read via mmap.

The Mmap object are cached to limit the system calls.

Methods

impl MmapDirectory

pub fn create_from_tempdir() -> Result<MmapDirectory>

Creates a new MmapDirectory in a temporary directory.

This is mostly useful to test the MmapDirectory itself. For your unit tests, prefer the RAMDirectory.

pub fn open<P: AsRef<Path>>(
directory_path: P
) -> Result<MmapDirectory, OpenDirectoryError>

Opens a MmapDirectory in a directory.

Returns an error if the directory_path does not exist or if it is not a directory.

pub fn get_cache_info(&mut self) -> CacheInfo

Returns some statistical information about the Mmap cache.

The MmapDirectory embeds a MmapDirectory to avoid multiplying the mmap system calls.