Trait tantivy::directory::TerminatingWrite

pub trait TerminatingWrite: Write + Send + Sync {
    fn terminate_ref(&mut self, AntiCallToken) -> Result<(), Error>;

    fn terminate(self) -> Result<(), Error> { ... }
}

Trait used to indicate when no more write need to be done on a writer

Required Methods

fn terminate_ref(&mut self, AntiCallToken) -> Result<(), Error>

You should implement this function to define custom behavior. This function should flush any buffer it may hold.

Provided Methods

fn terminate(self) -> Result<(), Error>

Indicate that the writer will no longer be used. Internally call terminate_ref.