Enum tantivy::Error

pub enum Error {
    PathDoesNotExist(PathBuf),
    FileAlreadyExists(PathBuf),
    IndexAlreadyExists,
    LockFailure(LockError, Option<String>),
    IOError(IOError),
    DataCorruption(DataCorruption),
    Poisoned,
    InvalidArgument(String),
    ErrorInThread(String),
    SchemaError(String),
    FastFieldError(FastFieldNotAvailableError),
    SystemError(String),
}

The library's failure based error enum

Variants
PathDoesNotExist(PathBuf)

Path does not exist.

FileAlreadyExists(PathBuf)

File already exists, this is a problem when we try to write into a new file.

IndexAlreadyExists

Index already exists in this directory

LockFailure(LockError, Option<String>)

Failed to acquire file lock

IOError(IOError)

IO Error.

DataCorruption(DataCorruption)

Data corruption.

Poisoned

A thread holding the locked panicked and poisoned the lock.

InvalidArgument(String)

Invalid argument was passed by the user.

ErrorInThread(String)

An Error happened in one of the thread.

SchemaError(String)

An Error appeared related to the schema.

FastFieldError(FastFieldNotAvailableError)

Tried to access a fastfield reader for a field not configured accordingly.

SystemError(String)

System error. (e.g.: We failed spawning a new thread)