Enum tantivy::ErrorKind

pub enum ErrorKind {
    Msg(String),
    PathDoesNotExist(PathBuf),
    FileAlreadyExists(PathBuf),
    IOError(IOError),
    CorruptedFile(PathBuf),
    Poisoned,
    InvalidArgument(String),
    ErrorInThread(String),
    SchemaError(String),
    FastFieldError(FastFieldNotAvailableError),
}

The kind of an error.

Variants
Msg(String)

A convenient variant for String.

PathDoesNotExist(PathBuf)

Path does not exist.

FileAlreadyExists(PathBuf)

File already exists, this is a problem when we try to write into a new file.

IOError(IOError)

IO Error.

CorruptedFile(PathBuf)

The data within is corrupted.

For instance, it contains invalid JSON.

Poisoned

A thread holding the locked panicked and poisoned the lock.

InvalidArgument(String)

Invalid argument was passed by the user.

ErrorInThread(String)

An Error happened in one of the thread.

SchemaError(String)

An Error appeared related to the lack of a field.

FastFieldError(FastFieldNotAvailableError)

Tried to access a fastfield reader for a field not configured accordingly.

Methods

impl ErrorKind

pub fn description(&self) -> &str

A string describing the error kind.