Enum tantivy::Executor

pub enum Executor {
    SingleThread,
    ThreadPool(ThreadPool),
}

Search executor whether search request are single thread or multithread.

We don’t expose Rayon thread pool directly here for several reasons.

First dependency hell. It is not a good idea to expose the API of a dependency, knowing it might conflict with a different version used by the client. Second, we may stop using rayon in the future.

Variants

SingleThread

Single thread variant of an Executor

ThreadPool(ThreadPool)

Thread pool variant of an Executor

Implementations

impl Executor

pub fn single_thread() -> Executor

Creates an Executor that performs all task in the caller thread.

pub fn multi_thread(num_threads: usize, prefix: &'static str) -> Result<Executor>

Creates an Executor that dispatches the tasks in a thread pool.

pub fn map<A: Send, R: Send, AIterator: Iterator<Item = A>, F: Sized + Sync + Fn(A) -> Result<R>>(
&self,
f: F,
args: AIterator
) -> Result<Vec<R>>

Perform a map in the thread pool.

Regardless of the executor (SingleThread or ThreadPool), panics in the task will propagate to the caller.