Enum tantivy::error::TantivyError

pub enum TantivyError {

    OpenDirectoryError(OpenDirectoryError),
    OpenReadError(OpenReadError),
    OpenWriteError(OpenWriteError),
    IndexAlreadyExists,
    LockFailure(LockError, Option<String>),
    IoError(Arc<Error>),
    DataCorruption(DataCorruption),
    Poisoned,
    FieldNotFound(String),
    InvalidArgument(String),
    ErrorInThread(String),
    IndexBuilderMissingArgument(&'static str),
    SchemaError(String),
    SystemError(String),
    IncompatibleIndex(Incompatibility),
    InternalError(String),

}

The library’s error enum

Variants

OpenDirectoryError(OpenDirectoryError)

Failed to open the directory.

OpenReadError(OpenReadError)

Failed to open a file for read.

OpenWriteError(OpenWriteError)

Failed to open a file for write.

IndexAlreadyExists

Index already exists in this directory.

LockFailure(LockError, Option<String>)

Failed to acquire file lock.

IoError(Arc<Error>)

IO Error.

DataCorruption(DataCorruption)

Data corruption.

Poisoned

A thread holding the locked panicked and poisoned the lock.

FieldNotFound(String)

The provided field name does not exist.

InvalidArgument(String)

Invalid argument was passed by the user.

ErrorInThread(String)

An Error occurred in one of the threads.

IndexBuilderMissingArgument(&'static str)

An Error occurred related to opening or creating a index.

SchemaError(String)

An Error occurred related to the schema.

SystemError(String)

System error. (e.g.: We failed spawning a new thread).

IncompatibleIndex(Incompatibility)

Index incompatible with current version of Tantivy.

InternalError(String)

An internal error occurred. This is are internal states that should not be reached. e.g. a datastructure is incorrectly inititalized.