Struct tantivy::fastfield::FastFieldReader

pub struct FastFieldReader<Item: FastValue> { /* fields omitted */ }

Trait for accessing a fastfield.

Depending on the field type, a different fast field is required.

Methods

impl<Item: FastValue> FastFieldReader<Item>

pub fn open(data: ReadOnlySource) -> Self

Opens a fast field given a source.

pub fn get(&self, doc: DocId) -> Item

Return the value associated to the given document.

This accessor should return as fast as possible.

Panics

May panic if doc is greater than the segment

pub fn get_range(&self, start: u32, output: &mut [Item])

Fills an output buffer with the fast field values associated with the DocId going from start to start + output.len().

Panics

May panic if start + output.len() is greater than the segment's maxdoc.

pub fn min_value(&self) -> Item

Returns the minimum value for this fast field.

The max value does not take in account of possible deleted document, and should be considered as an upper bound of the actual maximum value.

pub fn max_value(&self) -> Item

Returns the maximum value for this fast field.

The max value does not take in account of possible deleted document, and should be considered as an upper bound of the actual maximum value.