Struct tantivy::fastfield::FastFieldReaders

pub struct FastFieldReaders { /* private fields */ }

Provides access to all of the BitpackedFastFieldReader.

Internally, FastFieldReaders have preloaded fast field readers, and just wraps several HashMap.

Implementations

impl FastFieldReaders

pub fn u64(&self, field: Field) -> Result<Arc<dyn Column<u64>>>

Returns the u64 fast field reader reader associated with field.

If field is not a u64 fast field, this method returns an Error.

pub fn ip_addr(&self, field: Field) -> Result<Arc<dyn Column<Ipv6Addr>>>

Returns the ip fast field reader reader associated to field.

If field is not a u128 fast field, this method returns an Error.

pub fn ip_addrs(
&self,
field: Field
) -> Result<MultiValuedU128FastFieldReader<Ipv6Addr>>

Returns the ip fast field reader reader associated to field.

If field is not a u128 fast field, this method returns an Error.

pub fn u128s(&self, field: Field) -> Result<MultiValuedU128FastFieldReader<u128>>

Returns the u128 multi-valued fast field reader reader associated to field.

If field is not a u128 multi-valued fast field, this method returns an Error.

pub fn u64_lenient(&self, field: Field) -> Result<Arc<dyn Column<u64>>>

Returns the u64 fast field reader reader associated with field, regardless of whether the given field is effectively of type u64 or not.

If not, the fastfield reader will returns the u64-value associated with the original FastValue.

pub fn i64(&self, field: Field) -> Result<Arc<dyn Column<i64>>>

Returns the i64 fast field reader reader associated with field.

If field is not a i64 fast field, this method returns an Error.

pub fn date(&self, field: Field) -> Result<Arc<dyn Column<DateTime>>>

Returns the date fast field reader reader associated with field.

If field is not a date fast field, this method returns an Error.

pub fn f64(&self, field: Field) -> Result<Arc<dyn Column<f64>>>

Returns the f64 fast field reader reader associated with field.

If field is not a f64 fast field, this method returns an Error.

pub fn bool(&self, field: Field) -> Result<Arc<dyn Column<bool>>>

Returns the bool fast field reader reader associated with field.

If field is not a bool fast field, this method returns an Error.

pub fn u64s(&self, field: Field) -> Result<MultiValuedFastFieldReader<u64>>

Returns a u64s multi-valued fast field reader reader associated with field.

If field is not a u64 multi-valued fast field, this method returns an Error.

pub fn u64s_lenient(
&self,
field: Field
) -> Result<MultiValuedFastFieldReader<u64>>

Returns a u64s multi-valued fast field reader reader associated with field, regardless of whether the given field is effectively of type u64 or not.

If field is not a u64 multi-valued fast field, this method returns an Error.

pub fn i64s(&self, field: Field) -> Result<MultiValuedFastFieldReader<i64>>

Returns a i64s multi-valued fast field reader reader associated with field.

If field is not a i64 multi-valued fast field, this method returns an Error.

pub fn f64s(&self, field: Field) -> Result<MultiValuedFastFieldReader<f64>>

Returns a f64s multi-valued fast field reader reader associated with field.

If field is not a f64 multi-valued fast field, this method returns an Error.

pub fn bools(&self, field: Field) -> Result<MultiValuedFastFieldReader<bool>>

Returns a bools multi-valued fast field reader reader associated with field.

If field is not a bool multi-valued fast field, this method returns an Error.

pub fn dates(&self, field: Field) -> Result<MultiValuedFastFieldReader<DateTime>>

Returns a time::OffsetDateTime multi-valued fast field reader reader associated with field.

If field is not a time::OffsetDateTime multi-valued fast field, this method returns an Error.

pub fn bytes(&self, field: Field) -> Result<BytesFastFieldReader>

Returns the bytes fast field reader associated with field.

If field is not a bytes fast field, returns an Error.