Struct tantivy::fastfield::FastFieldSerializer

pub struct FastFieldSerializer { /* fields omitted */ }

FastFieldSerializer is in charge of serializing fastfields on disk.

Fast fields are encoded using bit-packing.

FastFieldWriters are in charge of pushing the data to the serializer. The serializer expects to receive the following calls.

new_u64_fast_field(...)

add_val(...)

add_val(...)

add_val(...)

...

close_field()

new_u64_fast_field(...)

add_val(...)

...

close_field()

close()

Methods

impl FastFieldSerializer

fn new(write: WritePtr) -> Result<FastFieldSerializer>

Constructor

fn new_u64_fast_field(
&mut self,
field: Field,
min_value: u64,
max_value: u64
) -> Result<()>

Start serializing a new u64 fast field

fn add_val(&mut self, val: u64) -> Result<()>

Pushes a new value to the currently open u64 fast field.

fn close_field(&mut self) -> Result<()>

Close the u64 fast field.

fn close(self) -> Result<usize>

Closes the serializer

After this call the data must be persistently save on disk.