Struct tantivy::fastfield::FastFieldsReader

pub struct FastFieldsReader { /* fields omitted */ }

The FastFieldsReader is the datastructure containing all of the fast fields' data.

It contains a mapping that associated these fields to the proper slice in the fastfield reader file.

Methods

impl FastFieldsReader

fn from_source(source: ReadOnlySource) -> Result<FastFieldsReader>

Opens a FastFieldsReader

When opening the fast field reader, the the list of the offset is read (as a footer of the data file).

fn open_reader<FFReader: FastFieldReader>(
&self,
field: Field
) -> Option<FFReader>

Returns the u64 fast value reader if the field is a u64 field indexed as "fast".

Return None if the field is not a u64 field indexed with the fast option.

Panics

May panic if the index is corrupted.