Struct tantivy::fastfield::FastFieldsWriter

pub struct FastFieldsWriter { /* private fields */ }

The FastFieldsWriter groups all of the fast field writers.

Implementations

impl FastFieldsWriter

pub fn from_schema(schema: &Schema) -> FastFieldsWriter

Create all FastFieldWriter required by the schema.

pub fn mem_usage(&self) -> usize

The memory used (inclusive childs)

pub fn get_term_id_writer(
&self,
field: Field
) -> Option<&MultiValuedFastFieldWriter>

Get the FastFieldWriter associated with a field.

pub fn get_field_writer(&self, field: Field) -> Option<&IntFastFieldWriter>

Get the FastFieldWriter associated with a field.

pub fn get_field_writer_mut(
&mut self,
field: Field
) -> Option<&mut IntFastFieldWriter>

Get the FastFieldWriter associated with a field.

pub fn get_term_id_writer_mut(
&mut self,
field: Field
) -> Option<&mut MultiValuedFastFieldWriter>

Get the FastFieldWriter associated with a field.

pub fn get_multivalue_writer_mut(
&mut self,
field: Field
) -> Option<&mut MultiValuedFastFieldWriter>

Returns the fast field multi-value writer for the given field.

Returns None if the field does not exist, or is not configured as a multivalued fastfield in the schema.

pub fn get_bytes_writer_mut(
&mut self,
field: Field
) -> Option<&mut BytesFastFieldWriter>

Returns the bytes fast field writer for the given field.

Returns None if the field does not exist, or is not configured as a bytes fastfield in the schema.

pub fn add_document(&mut self, doc: &Document) -> Result<()>

Indexes all of the fastfields of a new document.

pub fn serialize(
self,
serializer: &mut CompositeFastFieldSerializer,
mapping: &HashMap<Field, FxHashMap<u64, TermOrdinal>>,
doc_id_map: Option<&DocIdMapping>
) -> Result<()>

Serializes all of the FastFieldWriters by pushing them in order to the fast field serializer.