Trait tantivy::fastfield::FastValue

pub trait FastValue: MonotonicallyMappableToU64 + Copy + Send + Sync + PartialOrd + 'static {
    fn to_type() -> Type;

    fn make_zero() -> Self { ... }
}

Trait for types that are allowed for fast fields: (u64, i64 and f64, bool, DateTime).

Required Methods

fn to_type() -> Type

Returns the schema::Type for this FastValue.

Provided Methods

fn make_zero() -> Self

Build a default value. This default value is never used, so the value does not really matter.