Struct tantivy::fieldnorm::FieldNormReader

pub struct FieldNormReader { /* fields omitted */ }

Reads the fieldnorm associated to a document. The fieldnorm represents the length associated to a given Field of a given document.

This metric is important to compute the score of a document : a document having a query word in one its short fields (e.g. title) is likely to be more relevant than in one of its longer field (e.g. body).

tantivy encodes fieldnorm on one byte with some precision loss, using the same scheme as Lucene. Each value is place on a log-scale that takes values from 0 to 255.

A value on this scale is identified by a fieldnorm_id. Apart from compression, this scale also makes it possible to precompute computationally expensive functions of the fieldnorm in a very short array.

Methods

impl FieldNormReader

pub fn open(data: ReadOnlySource) -> Self

Opens a field norm reader given its data source.

pub fn fieldnorm(&self, doc_id: DocId) -> u32

Returns the fieldnorm associated to a doc id. The fieldnorm is a value approximating the number of tokens in a given field of the doc_id.

It is imprecise, and always lower than the actual number of tokens.

The fieldnorm is effectively decoded from the fieldnorm_id by doing a simple table lookup.

pub fn fieldnorm_id(&self, doc_id: DocId) -> u8

Returns the fieldnorm_id associated to a document.

pub fn id_to_fieldnorm(id: u8) -> u32

Converts a fieldnorm_id into a fieldnorm.

pub fn fieldnorm_to_id(fieldnorm: u32) -> u8

Converts a fieldnorm into a fieldnorm_id. (This function is not injective).