Struct tantivy::fieldnorm::FieldNormsWriter

pub struct FieldNormsWriter { /* private fields */ }

The FieldNormsWriter is in charge of tracking the fieldnorm byte of each document for each field with field norms.

FieldNormsWriter stores a Vec<u8> for each tracked field, using a byte per document per field.

Implementations

impl FieldNormsWriter

pub fn for_schema(schema: &Schema) -> FieldNormsWriter

Initialize with state for tracking the field norm fields specified in the schema.

pub fn mem_usage(&self) -> usize

The memory used inclusive childs

pub fn fill_up_to_max_doc(&mut self, max_doc: DocId)

Ensure that all documents in 0..max_doc have a byte associated with them in each of the fieldnorm vectors.

Will extend with 0-bytes for documents that have not been seen.

pub fn record(&mut self, doc: DocId, field: Field, fieldnorm: u32)

Set the fieldnorm byte for the given document for the given field.

Will internally convert the u32 fieldnorm value to the appropriate byte to approximate the field norm in less space.

doc - the document id

field - the field being set

fieldnorm - the number of terms present in document doc in field field

pub fn serialize(
&self,
fieldnorms_serializer: FieldNormsSerializer,
doc_id_map: Option<&DocIdMapping>
) -> Result<()>

Serialize the seen fieldnorm values to the serializer for all fields.