Function tantivy::demux

pub fn demux(
    segments: &[Segment],
    demux_mapping: &DemuxMapping,
    target_settings: IndexSettings,
    output_directories: Vec<Box<dyn Directory>>
) -> Result<Vec<Index>>

Demux the segments according to demux_mapping. See DemuxMapping. The number of output_directories need to match max new segment ordinal from demux_mapping.

The ordinal of segments need to match the ordinals provided in demux_mapping.