Module tantivy::postings

Postings module (also called inverted index)

Structs

BlockSegmentPostings

BlockSegmentPostings is a cursor iterating over blocks of documents.

FieldSerializer

The field serializer is in charge of the serialization of a specific field.

InvertedIndexSerializer

InvertedIndexSerializer is in charge of serializing postings on disk, in the

SegmentPostings

SegmentPostings represents the inverted list or postings associated with a term in a Segment.

TermInfo

TermInfo wraps the metadata associated with a Term. It is segment-local.

Traits

Postings

Postings (also called inverted list)