Enum tantivy::postings::SegmentPostingsOption

pub enum SegmentPostingsOption {
    NoFreq,
    Freq,
    FreqAndPositions,
}

Object describing the amount of information required when reading a postings.

Since decoding information is not free, this makes it possible to avoid this extra cost when the information is not required. For instance, positions are useful when running phrase queries but useless in other queries.

Variants
NoFreq

Only the doc ids are decoded

Freq

DocIds and term frequencies are decoded

FreqAndPositions

DocIds, term frequencies and positions will be decoded.