Enum tantivy::postings::SkipResult

pub enum SkipResult {
    Reached,
    OverStep,
    End,
}

Expresses the outcome of a call to DocSet's .skip_next(...).

Variants
Reached

target was in the docset

OverStep

target was not in the docset, skipping stopped as a greater element was found

End

the docset was entirely consumed without finding the target, nor any element greater than the target.