Struct tantivy::postings::BlockSegmentPostings

pub struct BlockSegmentPostings<'a> { /* fields omitted */ }

BlockSegmentPostings is a cursor iterating over blocks of documents.

Warning

While it is useful for some very specific high-performance use cases, you should prefer using SegmentPostings for most usage.

Methods

impl<'a> BlockSegmentPostings<'a>

fn doc_freq(&self) -> usize

Returns the document frequency associated to this block postings.

This doc_freq is simply the sum of the length of all of the blocks length, and it does not take in account deleted documents.

fn docs(&self) -> &[DocId]

Returns the array of docs in the current block.

Before the first call to .advance(), the block returned by .docs() is empty.

fn freq_handler(&self) -> &FreqHandler

Returns a reference to the frequency handler.

fn advance(&mut self) -> bool

Advance to the next block.

Returns false iff there was no remaining blocks.

fn empty() -> BlockSegmentPostings<'static>

Returns an empty segment postings object