Struct tantivy::postings::FreqHandler

pub struct FreqHandler { /* fields omitted */ }

FreqHandler is in charge of decompressing frequencies and/or positions.

Methods

impl FreqHandler

fn new_without_freq() -> FreqHandler

Returns a FreqHandler that just decodes DocIds.

fn new_with_freq() -> FreqHandler

Returns a FreqHandler that decodes DocIds and term frequencies.

fn new_with_freq_and_position(position_data: &[u8]) -> FreqHandler

Returns a FreqHandler that decodes DocIds, term frequencies, and term positions.

fn freq(&self, idx: usize) -> u32

Accessor to term frequency

idx is the offset of the current doc in the block. It takes value between 0 and 128.

fn positions(&self, idx: usize) -> &[u32]

Accessor to the positions

idx is the offset of the current doc in the block. It takes value between 0 and 128.

fn read_freq_block<'a>(&mut self, data: &'a [u8]) -> &'a [u8]

Decompresses a complete frequency block

fn read_freq_vint(&mut self, data: &[u8], num_els: usize)

Decompresses an incomplete frequency block