Struct tantivy::postings::IntersectionDocSet

pub struct IntersectionDocSet<TDocSet: DocSet> { /* fields omitted */ }

Creates a DocSet that iterator through the intersection of two DocSets.

Methods

impl<TDocSet: DocSet> IntersectionDocSet<TDocSet>

fn docsets(&self) -> &[TDocSet]

Returns an array to the underlying DocSets of the intersection. These DocSet are in the same position as the IntersectionDocSet, so that user can access their docfreq and positions.