Struct tantivy::postings::InvertedIndexSerializer

pub struct InvertedIndexSerializer { /* fields omitted */ }

InvertedIndexSerializer is in charge of serializing postings on disk, in the

.idx (inverted index)

.pos (positions file)

.term (term dictionary)

PostingsWriter are in charge of pushing the data to the serializer.

The serializer expects to receive the following calls in this order :

set_field(...)

new_term(...)

write_doc(...)

write_doc(...)

write_doc(...)

...

close_term()

new_term(...)

write_doc(...)

...

close_term()

set_field(...)

...

close()

Terms have to be pushed in a lexicographically-sorted order. Within a term, document have to be pushed in increasing order.

A description of the serialization format is available here.

Methods

impl InvertedIndexSerializer

pub fn open(segment: &mut Segment) -> Result<InvertedIndexSerializer>

Open a new PostingsSerializer for the given segment

pub fn new_field(
&mut self,
field: Field,
total_num_tokens: u64
) -> Result<FieldSerializer>

Must be called before starting pushing terms of a given field.

Loads the indexing options for the given field.

pub fn close(self) -> Result<()>

Closes the serializer.