Struct tantivy::postings::PostingsSerializer

pub struct PostingsSerializer { /* fields omitted */ }

PostingsSerializer is in charge of serializing postings on disk, in the * .idx (inverted index) * .pos (positions file) * .term (term dictionary)

PostingsWriter are in charge of pushing the data to the serializer.

The serializer expects to receive the following calls in this order :

new_term(...)

write_doc(...)

write_doc(...)

write_doc(...)

...

close_term()

new_term(...)

write_doc(...)

...

close_term()

close()

Terms have to be pushed in a lexicographically-sorted order. Within a term, document have to be pushed in increasing order.

A description of the serialization format is available here.

Methods

impl PostingsSerializer

fn new(terms_write: WritePtr,
postings_write: WritePtr,
positions_write: WritePtr,
schema: Schema)
-> Result<PostingsSerializer>

Open a new PostingsSerializer for the given segment

fn open(segment: &mut Segment) -> Result<PostingsSerializer>

Open a new PostingsSerializer for the given segment

fn new_term(&mut self, term: &Term, doc_freq: DocId) -> Result<()>

Starts the postings for a new term. * term - the term. It needs to come after the previous term according to the lexicographical order. * doc_freq - return the number of document containing the term.

fn close_term(&mut self) -> Result<()>

Finish the serialization for this term postings.

If the current block is incomplete, it need to be encoded using VInt encoding.

fn write_doc(&mut self,
doc_id: DocId,
term_freq: u32,
position_deltas: &[u32])
-> Result<()>

Serialize the information that a document contains the current term, its term frequency, and the position deltas.

At this point, the positions are already delta-encoded. For instance, if the positions are 2, 3, 17, position_deltas is 2, 1, 14

Term frequencies and positions may be ignored by the serializer depending on the configuration of the field in the Schema.

fn close(self) -> Result<()>

Closes the serializer.