Struct tantivy::postings::SegmentPostings

pub struct SegmentPostings { /* private fields */ }

SegmentPostings represents the inverted list or postings associated with a term in a Segment.

As we iterate through the SegmentPostings, the frequencies are optionally decoded. Positions on the other hand, are optionally entirely decoded upfront.

Implementations

impl SegmentPostings

pub fn empty() -> Self

Returns an empty segment postings object

pub fn doc_freq_given_deletes(&self, alive_bitset: &AliveBitSet) -> u32

Compute the number of non-deleted documents.

This method will clone and scan through the posting lists. (this is a rather expensive operation).

pub fn doc_freq(&self) -> u32

Returns the overall number of documents in the block postings. It does not take in account whether documents are deleted or not.