Trait tantivy::postings::PostingsWriter

pub trait PostingsWriter {
    fn suscribe(&mut self, doc: DocId, pos: u32, term: &Term, heap: &Heap);
    fn serialize(&self,
             serializer: &mut PostingsSerializer,
             heap: &Heap)
             -> Result<()>;
    fn close(&mut self, heap: &Heap);

    fn index_text<'a>(&mut self,
                  doc_id: DocId,
                  field: Field,
                  field_values: &[&'a FieldValue],
                  heap: &Heap)
                  -> u32 { ... }
}

The PostingsWriter is in charge of receiving documenting and building a Segment in anonymous memory.

PostingsWriter writes in a Heap.

Required Methods

fn suscribe(&mut self, doc: DocId, pos: u32, term: &Term, heap: &Heap)

Record that a document contains a term at a given position.

doc - the document id

pos - the term position (expressed in tokens)

term - the term

heap - heap used to store the postings informations as well as the terms in the hashmap.

fn serialize(&self,
serializer: &mut PostingsSerializer,
heap: &Heap)
-> Result<()>

Serializes the postings on disk. The actual serialization format is handled by the PostingsSerializer.

fn close(&mut self, heap: &Heap)

Closes all of the currently open Recorder's.

Provided Methods

fn index_text<'a>(&mut self,
doc_id: DocId,
field: Field,
field_values: &[&'a FieldValue],
heap: &Heap)
-> u32

Tokenize a text and suscribe all of its token.