Trait tantivy::postings::Recorder

pub trait Recorder: HeapAllocable {
    fn current_doc(&self) -> u32;
    fn new_doc(&mut self, doc: DocId, heap: &Heap);
    fn record_position(&mut self, position: u32, heap: &Heap);
    fn close_doc(&mut self, heap: &Heap);
    fn doc_freq(&self) -> u32;
    fn serialize(&self,
             self_addr: u32,
             serializer: &mut PostingsSerializer,
             heap: &Heap)
             -> Result<()>;
}

Recorder is in charge of recording relevant information about the presence of a term in a document.

Depending on the TextIndexingOptions associated to the field, the recorder may records * the document frequency * the document id * the term frequency * the term positions

Required Methods

fn current_doc(&self) -> u32

Returns the current document

fn new_doc(&mut self, doc: DocId, heap: &Heap)

Starts recording information about a new document This method shall only be called if the term is within the document.

fn record_position(&mut self, position: u32, heap: &Heap)

Record the position of a term. For each document, this method will be called term_freq times.

fn close_doc(&mut self, heap: &Heap)

Close the document. It will help record the term frequency.

fn doc_freq(&self) -> u32

Returns the number of document that have been seen so far

fn serialize(&self,
self_addr: u32,
serializer: &mut PostingsSerializer,
heap: &Heap)
-> Result<()>

Pushes the postings information to the serializer.