Enum tantivy::query::Occur

pub enum Occur {
    Should,
    Must,
    MustNot,
}

Defines whether a term in a query must be present, should be present or must not be present.

Variants

Should

For a given document to be considered for scoring, at least one of the terms with the Should or the Must Occur constraint must be within the document.

Must

Document without the term are excluded from the search.

MustNot

Document that contain the term are excluded from the search.

Implementations

impl Occur

pub fn compose(left: Occur, right: Occur) -> Occur

Compose two occur values.