Enum tantivy::query::QueryParserError

pub enum QueryParserError {

    SyntaxError(String),
    UnsupportedQuery(String),
    FieldDoesNotExist(String),
    ExpectedInt(ParseIntError),
    ExpectedBase64(DecodeError),
    ExpectedFloat(ParseFloatError),
    ExpectedBool(ParseBoolError),
    AllButQueryForbidden,
    NoDefaultFieldDeclared,
    FieldNotIndexed(String),
    FieldDoesNotHavePositionsIndexed(String),
    UnknownTokenizer {
        tokenizer: String,
        field: String,
    },
    RangeMustNotHavePhrase,
    DateFormatError(Parse),
    FacetFormatError(FacetParseError),
    IpFormatError(AddrParseError),

}

Possible error that may happen when parsing a query.

Variants

SyntaxError(String)

Error in the query syntax

UnsupportedQuery(String)

This query is unsupported.

FieldDoesNotExist(String)

The query references a field that is not in the schema

ExpectedInt(ParseIntError)

The query contains a term for a u64 or i64-field, but the value is neither.

ExpectedBase64(DecodeError)

The query contains a term for a bytes field, but the value is not valid base64.

ExpectedFloat(ParseFloatError)

The query contains a term for a f64-field, but the value is not a f64.

ExpectedBool(ParseBoolError)

The query contains a term for a bool field, but the value is not a bool.

AllButQueryForbidden

It is forbidden queries that are only “excluding”. (e.g. -title:pop)

NoDefaultFieldDeclared

If no default field is declared, running a query without any field specified is forbbidden.

FieldNotIndexed(String)

The field searched for is not declared as indexed in the schema.

FieldDoesNotHavePositionsIndexed(String)

A phrase query was requested for a field that does not have any positions indexed.

UnknownTokenizer

Fields

tokenizer: String

The name of the tokenizer

field: String

The field name

The tokenizer for the given field is unknown The two argument strings are the name of the field, the name of the tokenizer

RangeMustNotHavePhrase

The query contains a range query with a phrase as one of the bounds. Only terms can be used as bounds.

DateFormatError(Parse)

The format for the date field is not RFC 3339 compliant.

FacetFormatError(FacetParseError)

The format for the facet field is invalid.

IpFormatError(AddrParseError)

The format for the ip field is invalid.