Struct tantivy::query::Explanation

pub struct Explanation { /* private fields */ }

Object describing the score of a given document. It is organized in trees.

.to_pretty_json() can be useful to print out a human readable representation of this tree when debugging a given score.

Implementations

impl Explanation

pub fn new<T: ToString>(description: T, value: Score) -> Explanation

Creates a new explanation object.

pub fn value(&self) -> Score

Returns the value associated with the current node.

pub fn add_detail(&mut self, child_explanation: Explanation)

Add some detail, explaining some part of the current node formula.

Details are treated as child of the current node.

pub fn add_context(&mut self, context: String)

Adds some extra context to the explanation.

pub fn add_const<T: ToString>(&mut self, name: T, value: Score)

Shortcut for self.details.push(Explanation::new(name, value));

pub fn to_pretty_json(&self) -> String

Returns an indented json representation of the explanation tree for debug usage.