Struct tantivy::query::FuzzyTermQuery

pub struct FuzzyTermQuery { /* fields omitted */ }

A Fuzzy Query matches all of the documents containing a specific term that is within Levenshtein distance

#[macro_use]
extern crate tantivy;
use tantivy::schema::{Schema, TEXT};
use tantivy::{Index, Result, Term};
use tantivy::collector::{Count, TopDocs};
use tantivy::query::FuzzyTermQuery;

fn example() -> Result<()> {
    let mut schema_builder = Schema::builder();
    let title = schema_builder.add_text_field("title", TEXT);
    let schema = schema_builder.build();
    let index = Index::create_in_ram(schema);
    {
        let mut index_writer = index.writer(3_000_000)?;
        index_writer.add_document(doc!(
            title => "The Name of the Wind",
        ));
        index_writer.add_document(doc!(
            title => "The Diary of Muadib",
        ));
        index_writer.add_document(doc!(
            title => "A Dairy Cow",
        ));
        index_writer.add_document(doc!(
            title => "The Diary of a Young Girl",
        ));
        index_writer.commit().unwrap();
    }
    let reader = index.reader()?;
    let searcher = reader.searcher();

    {

        let term = Term::from_field_text(title, "Diary");
        let query = FuzzyTermQuery::new(term, 1, true);
        let (top_docs, count) = searcher.search(&query, &(TopDocs::with_limit(2), Count)).unwrap();
        assert_eq!(count, 2);
        assert_eq!(top_docs.len(), 2);
    }

    Ok(())
}

Methods

impl FuzzyTermQuery

pub fn new(
term: Term,
distance: u8,
transposition_cost_one: bool
) -> FuzzyTermQuery

Creates a new Fuzzy Query

pub fn new_prefix(
term: Term,
distance: u8,
transposition_cost_one: bool
) -> FuzzyTermQuery

Creates a new Fuzzy Query that treats transpositions as cost one rather than two