Struct tantivy::query::OccurFilter

pub struct OccurFilter { /* fields omitted */ }

An OccurFilter represents a filter over a bitset of at most 64 elements.

It wraps some simple bitmask to compute the filter rapidly.

Methods

impl OccurFilter

pub fn accept(&self, ord_set: u64) -> bool

Returns true if the bitset is matching the occur list.

pub fn new(occurs: &[Occur]) -> OccurFilter

Builds an OccurFilter from a list of Occur.