Struct tantivy::query::TermQuery

pub struct TermQuery { /* fields omitted */ }

A Term query matches all of the documents containing a specific term.

The score associated is defined as idf * sqrt(term_freq / field norm) in which : * idf - inverse document frequency. * term_freq - number of occurrences of the term in the field * field norm - number of tokens in the field.

Methods

impl TermQuery

fn new(term: Term, segment_postings_options: SegmentPostingsOption) -> TermQuery

Creates a new term query.

fn specialized_weight(&self, searcher: &Searcher) -> TermWeight

Returns a weight object.

While .weight(...) returns a boxed trait object, this method return a specific implementation. This is useful for optimization purpose.