Trait tantivy::query::ScoreCombiner

pub trait ScoreCombiner: Default + Clone + Send + Copy + 'static {
    fn update<TScorer: Scorer>(&mut self, scorer: &mut TScorer);
    fn clear(&mut self);
    fn score(&self) -> Score;
}

The ScoreCombiner trait defines how to compute an overall score given a list of scores.

Required Methods

fn update<TScorer: Scorer>(&mut self, scorer: &mut TScorer)

Aggregates the score combiner with the given scorer.

The ScoreCombiner may decide to call .scorer.score() or not.

fn clear(&mut self)

Clears the score combiner state back to its initial state.

fn score(&self) -> Score

Returns the aggregate score.