Trait tantivy::query::Scorer

pub trait Scorer: Downcast + DocSet + 'static {
    fn score(&mut self) -> Score;
}

Scored set of documents matching a query within a specific segment.

See Query.

Required Methods

fn score(&mut self) -> Score

Returns the score.

This method will perform a bit of computation and is not cached.

Implementations

impl dyn Scorer

pub fn is<__T: Scorer>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: Scorer>(self: Box<Self>) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T: Scorer>(self: Rc<Self>) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T: Scorer>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T: Scorer>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.