Enum tantivy::schema::IndexRecordOption

pub enum IndexRecordOption {
    Basic,
    WithFreqs,
    WithFreqsAndPositions,
}

IndexRecordOption describes an amount information associated with a given indexed field.

It is both used to:

describe in the schema the amount of information that should be retained during indexing (See TextFieldIndexing::set_index_option())

request that a given amount of information to be decoded as one goes through a posting list. (See InvertedIndexReader::read_postings())

Variants

Basic

records only the DocIds

WithFreqs

records the document ids as well as the term frequency. The term frequency can help giving better scoring of the documents.

WithFreqsAndPositions

records the document id, the term frequency and the positions of the occurrences in the document. Positions are required to run a PhraseQuery.

Implementations

impl IndexRecordOption

pub fn has_freq(self) -> bool

Returns true if this option includes encoding term frequencies.

pub fn has_positions(self) -> bool

Returns true if this option include encoding term positions.