Enum tantivy::schema::Value

pub enum Value {
    Str(String),
    PreTokStr(PreTokenizedString),
    U64(u64),
    I64(i64),
    F64(f64),
    Bool(bool),
    Date(DateTime),
    Facet(Facet),
    Bytes(Vec<u8>),
    JsonObject(Map<String, Value>),
    IpAddr(Ipv6Addr),
}

Value represents the value of a any field. It is an enum over all over all of the possible field type.

Variants

Str(String)

The str type is used for any text information.

PreTokStr(PreTokenizedString)

Pre-tokenized str type,

U64(u64)

Unsigned 64-bits Integer u64

I64(i64)

Signed 64-bits Integer i64

F64(f64)

64-bits Float f64

Bool(bool)

Bool value

Date(DateTime)

Date/time with microseconds precision

Facet(Facet)

Facet

Bytes(Vec<u8>)

Arbitrarily sized byte array

JsonObject(Map<String, Value>)

Json object value.

IpAddr(Ipv6Addr)

IpV6 Address. Internally there is no IpV4, it needs to be converted to Ipv6Addr.

Implementations

impl Value

pub fn as_text(&self) -> Option<&str>

Returns the text value, provided the value is of the Str type. (Returns None if the value is not of the Str type).

pub fn as_facet(&self) -> Option<&Facet>

Returns the facet value, provided the value is of the Facet type. (Returns None if the value is not of the Facet type).

pub fn tokenized_text(&self) -> Option<&PreTokenizedString>

Returns the tokenized text, provided the value is of the PreTokStr type. (Returns None if the value is not of the PreTokStr type.)

pub fn as_u64(&self) -> Option<u64>

Returns the u64-value, provided the value is of the U64 type. (Returns None if the value is not of the U64 type)

pub fn as_i64(&self) -> Option<i64>

Returns the i64-value, provided the value is of the I64 type.

Returns None if the value is not of type I64.

pub fn as_f64(&self) -> Option<f64>

Returns the f64-value, provided the value is of the F64 type.

Returns None if the value is not of type F64.

pub fn as_bool(&self) -> Option<bool>

Returns the bool value, provided the value is of the Bool type.

Returns None if the value is not of type Bool.

pub fn as_date(&self) -> Option<DateTime>

Returns the Date-value, provided the value is of the Date type.

Returns None if the value is not of type Date.

pub fn as_bytes(&self) -> Option<&[u8]>

Returns the Bytes-value, provided the value is of the Bytes type.

Returns None if the value is not of type Bytes.

pub fn as_json(&self) -> Option<&Map<String, Value>>

Returns the json object, provided the value is of the JsonObject type.

Returns None if the value is not of type JsonObject.

pub fn as_ip_addr(&self) -> Option<Ipv6Addr>

Returns the ip addr, provided the value is of the Ip type. (Returns None if the value is not of the Ip type)