Struct tantivy::schema::Document

pub struct Document { /* fields omitted */ }

Tantivy's Document is the object that can be indexed and then searched for.

Documents are fundamentally a collection of unordered couple (field, value). In this list, one field may appear more than once.

Documents are really just a list of couple (field, value). In this list, one field may appear more than once.

Methods

impl Document

pub fn new() -> Document

Creates a new, empty document object

pub fn len(&self) -> usize

Returns the number of (field, value) pairs.

pub fn is_empty(&self) -> bool

Returns true iff the document contains no fields.

pub fn filter_fields<P: Fn(Field) -> bool>(&mut self, predicate: P)

Retain only the field that are matching the predicate given in argument.

pub fn add_facet<F>(&mut self, field: Field, path: F) where
Facet: From<F>,

Adding a facet to the document.

pub fn add_text(&mut self, field: Field, text: &str)

Add a text field.

pub fn add_u64(&mut self, field: Field, value: u64)

Add a u64 field

pub fn add_i64(&mut self, field: Field, value: i64)

Add a u64 field

pub fn add(&mut self, field_value: FieldValue)

Add a field value

Important traits for &'a mut [u8]
impl<'a> Write for &'a mut [u8]impl<'a> Read for &'a [u8]

pub fn field_values(&self) -> &[FieldValue]

field_values accessor

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn get_sorted_field_values(&self) -> Vec<(Field, Vec<&FieldValue>)>

Sort and groups the field_values by field.

The result of this method is not cached and is computed on the fly when this method is called.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn get_all(&self, field: Field) -> Vec<&Value>

Returns all of the FieldValues associated the given field

pub fn get_first(&self, field: Field) -> Option<&Value>

Returns the first FieldValue associated the given field