Struct tantivy::schema::Facet

pub struct Facet(_);

A Facet represent a point in a given hierarchy.

They are typically represented similarly to a filepath. For instance, an e-commerce website could have a Facet for /electronics/tv_and_video/led_tv.

A document can be associated with any number of facets. The hierarchy implicitly imply that a document belonging to a facet also belongs to the ancestor of its facet. In the example above, /electronics/tv_and_video/ and /electronics.

Implementations

impl Facet

pub fn root() -> Facet

Returns a new instance of the “root facet” Equivalent to /.

pub fn is_root(&self) -> bool

Returns true if the facet is the root facet /.

pub fn encoded_str(&self) -> &str

Returns a binary representation of the facet.

In this representation, 0u8 is used as a separator and the string parts of the facet are unescaped. (The first / is not encoded at all).

This representation has the benefit of making it possible to express “being a child of a given facet” as a range over the term ordinals.

pub fn from_encoded(encoded_bytes: Vec<u8>) -> Result<Facet, FromUtf8Error>

Creates a Facet from its binary representation.

pub fn from_text<T>(path: &T) -> Result<Facet, FacetParseError>where
T: ?Sized + AsRef<str>,

Parse a text representation of a facet.

If one of the segments of this path contains a /, it should be escaped using an anti-slash \.

pub fn from_path<Path>(path: Path) -> Facetwhere
Path: IntoIterator,
Path::Item: ToString,

Returns a Facet from an iterator over the different steps of the facet path.

The steps are expected to be unescaped.

pub fn is_prefix_of(&self, other: &Facet) -> bool

Returns true if other is a strict subfacet of self.

Disclaimer: By strict we mean that the relation is not reflexive. /happy is not a prefix of /happy.

pub fn to_path(&self) -> Vec<&str>

Extract path from the Facet.

pub fn to_path_string(&self) -> String

This function is the inverse of Facet::from(&str).