Struct tantivy::schema::FieldEntry

pub struct FieldEntry { /* fields omitted */ }

A FieldEntry represents a field and its configuration. Schema are a collection of FieldEntry

It consists of

a field name

a field type, itself wrapping up options describing how the field should be indexed.

Methods

impl FieldEntry

pub fn new_text(field_name: String, text_options: TextOptions) -> FieldEntry

Creates a new u64 field entry in the schema, given a name, and some options.

pub fn new_u64(field_name: String, field_type: IntOptions) -> FieldEntry

Creates a new u64 field entry in the schema, given a name, and some options.

pub fn new_i64(field_name: String, field_type: IntOptions) -> FieldEntry

Creates a new i64 field entry in the schema, given a name, and some options.

pub fn new_facet(field_name: String) -> FieldEntry

Creates a field entry for a facet.

pub fn name(&self) -> &str

Returns the name of the field

pub fn field_type(&self) -> &FieldType

Returns the field type

pub fn is_indexed(&self) -> bool

Returns true iff the field is indexed

pub fn is_int_fast(&self) -> bool

Returns true iff the field is a int (signed or unsigned) fast field

pub fn is_stored(&self) -> bool

Returns true iff the field is stored