Struct tantivy::schema::IntOptions

pub struct IntOptions { /* fields omitted */ }

Define how an int field should be handled by tantivy.

Methods

impl IntOptions

pub fn is_stored(&self) -> bool

Returns true iff the value is stored.

pub fn is_indexed(&self) -> bool

Returns true iff the value is indexed.

pub fn is_fast(&self) -> bool

Returns true iff the value is a fast field.

pub fn set_stored(self) -> IntOptions

Set the u64 options as stored.

Only the fields that are set as stored are persisted into the Tantivy's store.

pub fn set_indexed(self) -> IntOptions

Set the u64 options as indexed.

Setting an integer as indexed will generate a posting list for each value taken by the integer.

pub fn set_fast(self, cardinality: Cardinality) -> IntOptions

Set the u64 options as a single-valued fast field.

Fast fields are designed for random access. Access time are similar to a random lookup in an array. If more than one value is associated to a fast field, only the last one is kept.

pub fn get_fastfield_cardinality(&self) -> Option<Cardinality>

Returns the cardinality of the fastfield.

If the field has not been declared as a fastfield, then the method returns None.