Struct tantivy::schema::Schema

pub struct Schema {
    // some fields omitted
}

Tantivy has a very strict schema. You need to specify in advance, whether a field is indexed or not, stored or not, and RAM-based or not.

This is done by creating a schema object, and setting up the fields one by one. It is for the moment impossible to remove fields.

Examples

use tantivy::schema::*;

  let mut schema = Schema::new();
  let id_field = schema.add_text_field("id", STRING);
  let title_field = schema.add_text_field("title", TEXT);
  let body_field = schema.add_text_field("body", TEXT);

Methods

impl Schema

fn new() -> Schema

Creates a new, empty schema.

fn get_field_entry(&self, field: Field) -> &FieldEntry

fn get_field_name(&self, field: Field) -> &String

fn fields(&self) -> &Vec<FieldEntry>

fn get_field(&self, field_name: &str) -> Option<Field>

Returns the field options associated with a given name.

Panics

Panics if the field name does not exist. It is meant as an helper for user who created and control the content of their schema.

If panicking is not an option for you, you may use get(&self, field_name: &str).

fn add_u32_field(&mut self, field_name_str: &str, field_options: U32Options) -> Field

Creates a new field. Return the associated field handle.

fn add_text_field(&mut self, field_name_str: &str, field_options: TextOptions) -> Field

fn to_named_doc(&self, doc: &Document) -> NamedFieldDocument

fn parse_document(&self, doc_json: &str) -> Result<Document, DocParsingError>

Build a document object from a json-object.