Struct tantivy::schema::SchemaBuilder

pub struct SchemaBuilder { /* fields omitted */ }

Tantivy has a very strict schema. You need to specify in advance whether a field is indexed or not, stored or not, and RAM-based or not.

This is done by creating a schema object, and setting up the fields one by one. It is for the moment impossible to remove fields.

Examples

use tantivy::schema::*;

let mut schema_builder = SchemaBuilder::default();
let id_field = schema_builder.add_text_field("id", STRING);
let title_field = schema_builder.add_text_field("title", TEXT);
let body_field = schema_builder.add_text_field("body", TEXT);
let schema = schema_builder.build();

Methods

impl SchemaBuilder

pub fn new() -> SchemaBuilder

Create a new SchemaBuilder

pub fn add_u64_field(
&mut self,
field_name_str: &str,
field_options: IntOptions
) -> Field

Adds a new u64 field. Returns the associated field handle

Caution

Appending two fields with the same name will result in the shadowing of the first by the second one. The first field will get a field id but only the second one will be indexed

pub fn add_i64_field(
&mut self,
field_name_str: &str,
field_options: IntOptions
) -> Field

Adds a new i64 field. Returns the associated field handle

Caution

Appending two fields with the same name will result in the shadowing of the first by the second one. The first field will get a field id but only the second one will be indexed

pub fn add_text_field(
&mut self,
field_name_str: &str,
field_options: TextOptions
) -> Field

Adds a new text field. Returns the associated field handle

Caution

Appending two fields with the same name will result in the shadowing of the first by the second one. The first field will get a field id but only the second one will be indexed

pub fn add_facet_field(&mut self, field_name: &str) -> Field

Adds a facet field to the schema.

pub fn build(self) -> Schema

Finalize the creation of a Schema This will consume your SchemaBuilder