Struct tantivy::schema::Term

pub struct Term<B = Vec<u8>>(_)
 where
    B: AsRef<[u8]>;

Term represents the value that the token can take.

It actually wraps a Vec<u8>.

Methods

impl Term

pub fn from_field_i64(field: Field, val: i64) -> Term

Builds a term given a field, and a u64-value

Assuming the term has a field id of 1, and a u64 value of 3234, the Term will have 8 bytes.

The first four byte are dedicated to storing the field id as a u64. The 4 following bytes are encoding the u64 value.

pub fn from_field_date(field: Field, val: &DateTime) -> Term

Builds a term given a field, and a DateTime value

Assuming the term has a field id of 1, and a timestamp i64 value of 3234, the Term will have 8 bytes.

The first four byte are dedicated to storing the field id as a u64. The 4 following bytes are encoding the DateTime as i64 timestamp value.

pub fn from_facet(field: Field, facet: &Facet) -> Term

Creates a Term given a facet.

pub fn from_field_text(field: Field, text: &str) -> Term

Builds a term given a field, and a string value

Assuming the term has a field id of 2, and a text value of "abc", the Term will have 4 bytes. The first byte is 2, and the three following bytes are the utf-8 representation of "abc".

pub fn from_field_u64(field: Field, val: u64) -> Term

Builds a term given a field, and a u64-value

Assuming the term has a field id of 1, and a u64 value of 3234, the Term will have 8 bytes.

The first four byte are dedicated to storing the field id as a u64. The 4 following bytes are encoding the u64 value.

pub fn set_field(&mut self, field: Field)

Returns the field.

pub fn set_u64(&mut self, val: u64)

Sets a u64 value in the term.

U64 are serialized using (8-byte) BigEndian representation. The use of BigEndian has the benefit of preserving the natural order of the values.

pub fn set_i64(&mut self, val: i64)

Sets a i64 value in the term.

pub fn set_text(&mut self, text: &str)

Set the texts only, keeping the field untouched.

impl<B> Term<B> where
B: AsRef<[u8]>,

pub fn wrap(data: B) -> Term<B>

Wraps a source of data

pub fn field(&self) -> Field

Returns the field.

pub fn get_u64(&self) -> u64

Returns the u64 value stored in a term.

Panics

... or returns an invalid value if the term is not a u64 field.

pub fn get_i64(&self) -> i64

Returns the i64 value stored in a term.

Panics

... or returns an invalid value if the term is not a i64 field.

pub fn text(&self) -> &str

Returns the text associated with the term.

Panics

If the value is not valid utf-8. This may happen if the index is corrupted or if you try to call this method on a non-string type.

pub fn value_bytes(&self) -> &[u8]

Returns the serialized value of the term. (this does not include the field.)

If the term is a string, its value is utf-8 encoded. If the term is a u64, its value is encoded according to byteorder::LittleEndian.

pub fn as_slice(&self) -> &[u8]

Returns the underlying &[u8]