Struct tantivy::space_usage::FieldUsage

pub struct FieldUsage { /* private fields */ }

Represents space usage of a given field, breaking it down into the (field, index) pairs that comprise it.

See documentation for PerFieldSpaceUsage for slightly more information.

Implementations

impl FieldUsage

pub fn field(&self) -> Field

Field

pub fn sub_num_bytes(&self) -> &[Option<ByteCount>]

Space usage for each index

pub fn total(&self) -> ByteCount

Total bytes used for this field in this context