Struct tantivy::store::StoreWriter

pub struct StoreWriter { /* fields omitted */ }

Write tantivy's Store

Contrary to the other components of tantivy, the store is written to disc as document as being added, as opposed to when the segment is getting finalized.

The skip list index on the other hand, is build in memory.

Methods

impl StoreWriter

pub fn new(writer: WritePtr) -> StoreWriter

Create a store writer.

The store writer will writes blocks on disc as document are added.

pub fn store(&mut self, stored_document: &Document) -> Result<()>

Store a new document.

The document id is implicitely the number of times this method has been called.

pub fn stack(&mut self, store_reader: &StoreReader) -> Result<()>

Stacks a store reader on top of the documents written so far. This method is an optimization compared to iterating over the documents in the store and adding them one by one, as the store's data will not be decompressed and then recompressed.

pub fn close(self) -> Result<()>

Finalized the store writer.

Compress the last unfinished block if any, and serializes the skip list index on disc.