Struct tantivy::DateOptions

pub struct DateOptions { /* private fields */ }

Defines how DateTime field should be handled by tantivy.

Implementations

impl DateOptions

pub fn is_stored(&self) -> bool

Returns true iff the value is stored.

pub fn is_indexed(&self) -> bool

Returns true iff the value is indexed and therefore searchable.

pub fn fieldnorms(&self) -> bool

Returns true iff the field has fieldnorm.

pub fn is_multivalue_fast(&self) -> bool

Returns true iff the value is a fast field and multivalue.

pub fn is_fast(&self) -> bool

Returns true iff the value is a fast field.

pub fn set_stored(self) -> DateOptions

Set the field as stored.

Only the fields that are set as stored are persisted into the Tantivy’s store.

pub fn set_indexed(self) -> DateOptions

Set the field as indexed.

Setting an integer as indexed will generate a posting list for each value taken by the integer.

This is required for the field to be searchable.

pub fn set_fieldnorm(self) -> DateOptions

Set the field with fieldnorm.

Setting an integer as fieldnorm will generate the fieldnorm data for it.

pub fn set_fast(self, cardinality: Cardinality) -> DateOptions

Set the field as a single-valued fast field.

Fast fields are designed for random access. Access time are similar to a random lookup in an array. If more than one value is associated with a fast field, only the last one is kept.

pub fn get_fastfield_cardinality(&self) -> Option<Cardinality>

Returns the cardinality of the fastfield.

If the field has not been declared as a fastfield, then the method returns None.

pub fn set_precision(self, precision: DatePrecision) -> DateOptions

Sets the precision for this DateTime field.

Internal storage precision, used to optimize storage compression on fast fields.

pub fn get_precision(&self) -> DatePrecision

Returns the storage precision for this DateTime field.

Internal storage precision, used to optimize storage compression on fast fields.