Struct tantivy::Document

pub struct Document { /* private fields */ }

Tantivy’s Document is the object that can be indexed and then searched for.

Documents are fundamentally a collection of unordered couples (field, value). In this list, one field may appear more than once.

Implementations

impl Document

pub fn new() -> Document

Creates a new, empty document object

pub fn len(&self) -> usize

Returns the number of (field, value) pairs.

pub fn is_empty(&self) -> bool

Returns true if the document contains no fields.

pub fn add_facet<F>(&mut self, field: Field, path: F)where
Facet: From<F>,

Adding a facet to the document.

pub fn add_text<S: ToString>(&mut self, field: Field, text: S)

Add a text field.

pub fn add_pre_tokenized_text(
&mut self,
field: Field,
pre_tokenized_text: PreTokenizedString
)

Add a pre-tokenized text field.

pub fn add_u64(&mut self, field: Field, value: u64)

Add a u64 field

pub fn add_ip_addr(&mut self, field: Field, value: Ipv6Addr)

Add a IP address field. Internally only Ipv6Addr is used.

pub fn add_i64(&mut self, field: Field, value: i64)

Add a i64 field

pub fn add_f64(&mut self, field: Field, value: f64)

Add a f64 field

pub fn add_bool(&mut self, field: Field, value: bool)

Add a bool field

pub fn add_date(&mut self, field: Field, value: DateTime)

Add a date field with unspecified time zone offset

pub fn add_bytes<T: Into<Vec<u8>>>(&mut self, field: Field, value: T)

Add a bytes field

pub fn add_json_object(&mut self, field: Field, json_object: Map<String, Value>)

Add a JSON field

pub fn add_field_value<T: Into<Value>>(&mut self, field: Field, typed_val: T)

Add a (field, value) to the document.

pub fn field_values(&self) -> &[FieldValue]

field_values accessor

pub fn get_sorted_field_values(&self) -> Vec<(Field, Vec<&Value>)>

Sort and groups the field_values by field.

The result of this method is not cached and is computed on the fly when this method is called.

pub fn get_all(&self, field: Field) -> impl Iterator<Item = &Value>

Returns all of the FieldValues associated the given field

pub fn get_first(&self, field: Field) -> Option<&Value>

Returns the first FieldValue associated the given field

pub fn serialize_stored<W: Write>(
&self,
schema: &Schema,
writer: &mut W
) -> Result<()>

Serializes stored field values.