Struct tantivy::InvertedIndexReader

pub struct InvertedIndexReader { /* private fields */ }

The inverted index reader is in charge of accessing the inverted index associated with a specific field.

Note

It is safe to delete the segment associated with an InvertedIndexReader. As long as it is open, the FileSlice it is relying on should stay available.

InvertedIndexReader are created by calling SegmentReader::inverted_index().

Implementations

impl InvertedIndexReader

pub fn empty(record_option: IndexRecordOption) -> InvertedIndexReader

Creates an empty InvertedIndexReader object, which contains no terms at all.

pub fn get_term_info(&self, term: &Term) -> Result<Option<TermInfo>>

Returns the term info associated with the term.

pub fn terms(&self) -> &TermDictionary

Return the term dictionary datastructure.

pub fn reset_block_postings_from_terminfo( &self, term_info: &TermInfo, block_postings: &mut BlockSegmentPostings ) -> Result<()>

Resets the block segment to another position of the postings file.

This is useful for enumerating through a list of terms, and consuming the associated posting lists while avoiding reallocating a BlockSegmentPostings.

Warning

This does not reset the positions list.

pub fn read_block_postings( &self, term: &Term, option: IndexRecordOption ) -> Result<Option<BlockSegmentPostings>>

Returns a block postings given a Term. This method is for an advanced usage only.

Most users should prefer using Self::read_postings() instead.

pub fn read_block_postings_from_terminfo( &self, term_info: &TermInfo, requested_option: IndexRecordOption ) -> Result<BlockSegmentPostings>

Returns a block postings given a term_info. This method is for an advanced usage only.

Most users should prefer using Self::read_postings() instead.

pub fn read_postings_from_terminfo( &self, term_info: &TermInfo, option: IndexRecordOption ) -> Result<SegmentPostings>

Returns a posting object given a term_info. This method is for an advanced usage only.

Most users should prefer using Self::read_postings() instead.

pub fn total_num_tokens(&self) -> u64

Returns the total number of tokens recorded for all documents (including deleted documents).

pub fn read_postings( &self, term: &Term, option: IndexRecordOption ) -> Result<Option<SegmentPostings>>

Returns the segment postings associated with the term, and with the given option, or None if the term has never been encountered and indexed.

If the field was not indexed with the indexing options that cover the requested options, the returned SegmentPostings the method does not fail and returns a SegmentPostings with as much information as possible.

For instance, requesting IndexRecordOption::WithFreqs for a TextOptions that does not index position will return a SegmentPostings with DocIds and frequencies.

pub fn doc_freq(&self, term: &Term) -> Result<u32>

Returns the number of documents containing the term.