Struct tantivy::Searcher

pub struct Searcher { /* fields omitted */ }

Holds a list of SegmentReaders ready for search.

It guarantees that the Segment will not be removed before
the destruction of the Searcher.

Methods

impl Searcher

fn doc(&self, doc_address: &DocAddress) -> Result<Document>

Fetches a document from tantivy's store given a DocAddress.

The searcher uses the segment ordinal to route the the request to the right Segment.

fn num_docs(&self) -> DocId

Returns the overall number of documents in the index.

fn doc_freq(&self, term: &Term) -> u32

Return the overall number of documents containing the given term.

fn terms<'a>(&'a self) -> TermIterator<'a>

Returns a Stream over all of the sorted unique terms of the searcher.

This includes all of the fields from all of the segment_readers. See TermIterator.

Warning

This API is very likely to change in the future.

fn segment_readers(&self) -> &[SegmentReader]

Return the list of segment readers

fn segment_reader(&self, segment_ord: usize) -> &SegmentReader

Returns the segment_reader associated with the given segment_ordinal

fn search<C: Collector>(&self,
query: &Query,
collector: &mut C)
-> Result<TimerTree>

Runs a query on the segment readers wrapped by the searcher