Struct tantivy::Segment

pub struct Segment { /* fields omitted */ }

A segment is a piece of the index.

Methods

impl Segment

pub fn index(&self) -> &Index

Returns the index the segment belongs to.

pub fn schema(&self) -> Schema

Returns our index's schema.

pub fn meta(&self) -> &SegmentMeta

Returns the segment meta-information

pub fn id(&self) -> SegmentId

Returns the segment's id.

pub fn relative_path(&self, component: SegmentComponent) -> PathBuf

Returns the relative path of a component of our segment.

It just joins the segment id with the extension associated to a segment component.

pub fn protect_from_delete(&self, component: SegmentComponent) -> FileProtection

Protects a specific component file from being deleted.

Returns a FileProtection object. The file is guaranteed to not be garbage collected as long as this FileProtection object lives.

pub fn open_read(
&self,
component: SegmentComponent
) -> Result<ReadOnlySource, OpenReadError>

Open one of the component file for a regular read.

pub fn open_write(
&mut self,
component: SegmentComponent
) -> Result<WritePtr, OpenWriteError>

Open one of the component file for regular write.