Struct tantivy::SegmentMeta

pub struct SegmentMeta { /* fields omitted */ }

SegmentMeta contains simple meta information about a segment.

For instance the number of docs it contains, how many are deleted, etc.

Methods

impl SegmentMeta

pub fn new(segment_id: SegmentId) -> SegmentMeta

Creates a new segment meta for a segment with no deletes and no documents.

pub fn id(&self) -> SegmentId

Returns the segment id.

pub fn num_deleted_docs(&self) -> u32

Returns the number of deleted documents.

pub fn list_files(&self) -> HashSet<PathBuf>

Returns the list of files that are required for the segment meta.

This is useful as the way tantivy removes files is by removing all files that have been created by tantivy and are not used by any segment anymore.

pub fn relative_path(&self, component: SegmentComponent) -> PathBuf

Returns the relative path of a component of our segment.

It just joins the segment id with the extension associated to a segment component.

pub fn max_doc(&self) -> u32

Return the highest doc id + 1

If there are no deletes, then num_docs = max_docs and all the doc ids contains in this segment are exactly (0..max_doc).

pub fn num_docs(&self) -> u32

Return the number of documents in the segment.

pub fn delete_opstamp(&self) -> Option<u64>

Returns the opstamp of the last delete operation taken in account in this segment.

pub fn has_deletes(&self) -> bool

Returns true iff the segment meta contains delete information.