Struct tantivy::SegmentReader

pub struct SegmentReader { /* private fields */ }

Entry point to access all of the datastructures of the Segment

term dictionary

postings

store

fast field readers

field norm reader

The segment reader has a very low memory footprint, as close to all of the memory data is mmapped.

Implementations

impl SegmentReader

pub fn max_doc(&self) -> DocId

Returns the highest document id ever attributed in this segment + 1.

pub fn num_docs(&self) -> DocId

Returns the number of alive documents. Deleted documents are not counted.

pub fn schema(&self) -> &Schema

Returns the schema of the index this segment belongs to.

pub fn num_deleted_docs(&self) -> DocId

Return the number of documents that have been deleted in the segment.

pub fn has_deletes(&self) -> bool

Returns true if some of the documents of the segment have been deleted.

pub fn fast_fields(&self) -> &FastFieldReaders

Accessor to a segment’s fast field reader given a field.

Returns the u64 fast value reader if the field is a u64 field indexed as “fast”.

Return a FastFieldNotAvailableError if the field is not declared as a fast field in the schema.

Panics

May panic if the index is corrupted.

pub fn facet_reader(&self, field: Field) -> Result<FacetReader>

Accessor to the FacetReader associated with a given Field.

pub fn get_fieldnorms_reader(&self, field: Field) -> Result<FieldNormReader>

Accessor to the segment’s Field norms’s reader.

Field norms are the length (in tokens) of the fields. It is used in the computation of the TfIdf.

They are simply stored as a fast field, serialized in the .fieldnorm file of the segment.

pub fn get_store_reader(&self, cache_size: usize) -> Result<StoreReader>

Accessor to the segment’s StoreReader.

pub fn open(segment: &Segment) -> Result<SegmentReader>

Open a new segment for reading.

pub fn open_with_custom_alive_set(
segment: &Segment,
custom_bitset: Option<AliveBitSet>
) -> Result<SegmentReader>

Open a new segment for reading.

pub fn inverted_index(&self, field: Field) -> Result<Arc<InvertedIndexReader>>

Returns a field reader associated with the field given in argument. If the field was not present in the index during indexing time, the InvertedIndexReader is empty.

The field reader is in charge of iterating through the term dictionary associated with a specific field, and opening the posting list associated with any term.

If the field is not marked as index, a warning is logged and an empty InvertedIndexReader is returned. Similarly, if the field is marked as indexed but no term has been indexed for the given index, an empty InvertedIndexReader is returned (but no warning is logged).

pub fn segment_id(&self) -> SegmentId

Returns the segment id

pub fn delete_opstamp(&self) -> Option<Opstamp>

Returns the delete opstamp

pub fn alive_bitset(&self) -> Option<&AliveBitSet>

Returns the bitset representing the alive DocIds.

pub fn is_deleted(&self, doc: DocId) -> bool

Returns true if the doc is marked as deleted.

pub fn doc_ids_alive(&self) -> Box<dyn Iterator<Item = DocId> + '_>

Returns an iterator that will iterate over the alive document ids

pub fn space_usage(&self) -> Result<SegmentSpaceUsage>

Summarize total space usage of this segment.