Struct tantivy::Term

pub struct Term<B = Vec<u8>>(_)
where
    B: AsRef<[u8]>;

Term represents the value that the token can take.

It actually wraps a Vec<u8>.

Implementations

impl Term

pub fn is_empty(&self) -> bool

Is empty if there are no value bytes.

pub fn from_field_ip_addr(field: Field, ip_addr: Ipv6Addr) -> Term

Builds a term given a field, and a Ipv6Addr-value

pub fn from_field_u64(field: Field, val: u64) -> Term

Builds a term given a field, and a u64-value

pub fn from_field_i64(field: Field, val: i64) -> Term

Builds a term given a field, and a i64-value

pub fn from_field_f64(field: Field, val: f64) -> Term

Builds a term given a field, and a f64-value

pub fn from_field_bool(field: Field, val: bool) -> Term

Builds a term given a field, and a bool-value

pub fn from_field_date(field: Field, val: DateTime) -> Term

Builds a term given a field, and a DateTime value

pub fn from_facet(field: Field, facet: &Facet) -> Term

Creates a Term given a facet.

pub fn from_field_text(field: Field, text: &str) -> Term

Builds a term given a field, and a string value

pub fn from_field_bytes(field: Field, bytes: &[u8]) -> Term

Builds a term bytes.

pub fn clear_with_type(&mut self, typ: Type)

Removes the value_bytes and set the type code.

pub fn set_u64(&mut self, val: u64)

Sets a u64 value in the term.

U64 are serialized using (8-byte) BigEndian representation. The use of BigEndian has the benefit of preserving the natural order of the values.

pub fn set_i64(&mut self, val: i64)

Sets a i64 value in the term.

pub fn set_date(&mut self, date: DateTime)

Sets a DateTime value in the term.

pub fn set_f64(&mut self, val: f64)

Sets a f64 value in the term.

pub fn set_bool(&mut self, val: bool)

Sets a bool value in the term.

pub fn set_ip_addr(&mut self, val: Ipv6Addr)

Sets a Ipv6Addr value in the term.

pub fn set_bytes(&mut self, bytes: &[u8])

Sets the value of a Bytes field.

pub fn set_text(&mut self, text: &str)

Set the texts only, keeping the field untouched.

pub fn truncate_value_bytes(&mut self, len: usize)

Truncates the value bytes of the term. Value and field type stays the same.

pub fn value_bytes_mut(&mut self) -> &mut [u8]

Returns the value bytes as mutable slice

pub fn len_bytes(&self) -> usize

The length of the bytes.

pub fn append_bytes(&mut self, bytes: &[u8]) -> &mut [u8]

Appends value bytes to the Term.

This function returns the segment that has just been added.

pub fn push_byte(&mut self, byte: u8)

Appends a single byte to the term.

impl<B> Term<B>where
B: AsRef<[u8]>,

pub fn wrap(data: B) -> Term<B>

Wraps a object holding bytes

pub fn typ(&self) -> Type

Return the type of the term.

pub fn field(&self) -> Field

Returns the field.

pub fn as_u64(&self) -> Option<u64>

Returns the u64 value stored in a term.

Returns None if the term is not of the u64 type, or if the term byte representation is invalid.

pub fn as_i64(&self) -> Option<i64>

Returns the i64 value stored in a term.

Returns None if the term is not of the i64 type, or if the term byte representation is invalid.

pub fn as_f64(&self) -> Option<f64>

Returns the f64 value stored in a term.

Returns None if the term is not of the f64 type, or if the term byte representation is invalid.

pub fn as_bool(&self) -> Option<bool>

Returns the bool value stored in a term.

Returns None if the term is not of the bool type, or if the term byte representation is invalid.

pub fn as_date(&self) -> Option<DateTime>

Returns the Date value stored in a term.

Returns None if the term is not of the Date type, or if the term byte representation is invalid.

pub fn as_str(&self) -> Option<&str>

Returns the text associated with the term.

Returns None if the field is not of string type or if the bytes are not valid utf-8.

pub fn as_facet(&self) -> Option<Facet>

Returns the facet associated with the term.

Returns None if the field is not of facet type or if the bytes are not valid utf-8.

pub fn as_bytes(&self) -> Option<&[u8]>

Returns the bytes associated with the term.

Returns None if the field is not of bytes type.

pub fn value_bytes(&self) -> &[u8]

Returns the serialized value of the term. (this does not include the field.)

If the term is a string, its value is utf-8 encoded. If the term is a u64, its value is encoded according to byteorder::LittleEndian.