Struct tantivy::TermIterator

pub struct TermIterator<'a> { /* fields omitted */ }

Given a list of sorted term streams, returns an iterator over sorted unique terms.

The item yield is actually a pair with - the term - a slice with the ordinal of the segments containing the terms.

Methods

impl<'a> TermIterator<'a>

fn advance(&mut self) -> bool

Advance the term iterator to the next term. Returns true if there is indeed another term False if there is none.

fn term(&self) -> &Term

Returns the current term.

This method may be called iff advance() has been called before and "true" was returned.

fn segment_ords(&self) -> &[usize]

Returns the sorted list of segment ordinals that include the current term.

This method may be called iff advance() has been called before and "true" was returned.