Struct tantivy::TrackedObject

pub struct TrackedObject<T> { /* private fields */ }

Your tracked object.

A tracked object contains reference counting logic and an Arc to your object. It is cloneable but calling clone will not clone your internal object.

Your object cannot be mutated. You can borrow it using the Deref interface.

Implementations

impl<T> TrackedObject<T>

pub fn map<F>(&self, f: F) -> TrackedObject<T>where
F: FnOnce(&T) -> T,

Creates a new object from an existing one.

The new object will be registered in your original object’s inventory.

use census::{Inventory, TrackedObject};

let inventory = Inventory::new();

let seven = inventory.track(7);
let fourteen = seven.map(|i| i * 2);
assert_eq!(*fourteen, 14);

let living_instances = inventory.list();
assert_eq!(living_instances.len(), 2);