Struct tantivy::termdict::TermDictionary

pub struct TermDictionary { /* private fields */ }

The term dictionary contains all of the terms in tantivy index in a sorted manner.

The Fst crate is used to associate terms to their respective TermOrdinal. The TermInfoStore then makes it possible to fetch the associated TermInfo.

Implementations

impl TermDictionary

pub fn open(file: FileSlice) -> Result<Self>

Opens a TermDictionary.

pub fn empty() -> Self

Creates an empty term dictionary which contains no terms.

pub fn num_terms(&self) -> usize

Returns the number of terms in the dictionary. Term ordinals range from 0 to num_terms() - 1.

pub fn term_ord<K: AsRef<[u8]>>(&self, key: K) -> Result<Option<TermOrdinal>>

Returns the ordinal associated with a given term.

pub fn ord_to_term(&self, ord: TermOrdinal, bytes: &mut Vec<u8>) -> Result<bool>

Stores the term associated with a given term ordinal in a bytes buffer.

Term ordinals are defined as the position of the term in the sorted list of terms.

Returns true if and only if the term has been found.

Regardless of whether the term is found or not, the buffer may be modified.

pub fn term_info_from_ord(&self, term_ord: TermOrdinal) -> TermInfo

Returns the number of terms in the dictionary.

pub fn get<K: AsRef<[u8]>>(&self, key: K) -> Result<Option<TermInfo>>

Lookups the value corresponding to the key.

pub fn range(&self) -> TermStreamerBuilder<'_>

Returns a range builder, to stream all of the terms within an interval.

pub fn stream(&self) -> Result<TermStreamer<'_>>

A stream of all the sorted terms.

pub fn search<'a, A: Automaton + 'a>(
&'a self,
automaton: A
) -> TermStreamerBuilder<'a, A>

Returns a search builder, to stream all of the terms within the Automaton