Struct tantivy::termdict::TermMerger

pub struct TermMerger<'a, V> where
    V: 'a + BinarySerializable + Default,  { /* fields omitted */ }

Given a list of sorted term streams, returns an iterator over sorted unique terms.

The item yield is actually a pair with - the term - a slice with the ordinal of the segments containing the terms.

Methods

impl<'a, V> TermMerger<'a, V> where
V: 'a + BinarySerializable + Default,

fn advance(&mut self) -> bool

Advance the term iterator to the next term. Returns true if there is indeed another term False if there is none.

fn key(&self) -> &[u8]

Returns the current term.

This method may be called iff advance() has been called before and "true" was returned.

fn current_kvs(&self) -> &[HeapItem<'a, V>]

Returns the sorted list of segment ordinals that include the current term.

This method may be called iff advance() has been called before and "true" was returned.

fn next(&mut self) -> Option<Term<&[u8]>>

Iterates through terms