Trait tantivy::termdict::TermStreamerBuilder

pub trait TermStreamerBuilder {
    type Streamer: TermStreamer;
    fn ge<T: AsRef<[u8]>>(self, bound: T) -> Self;

    fn gt<T: AsRef<[u8]>>(self, bound: T) -> Self;

    fn lt<T: AsRef<[u8]>>(self, bound: T) -> Self;

    fn le<T: AsRef<[u8]>>(self, bound: T) -> Self;

    fn into_stream(self) -> Self::Streamer;
}

TermStreamerBuilder is an helper object used to define a range of terms that should be streamed.

Associated Types

type Streamer: TermStreamer

Associated TermStreamer type that this builder is building.

Required Methods

fn ge<T: AsRef<[u8]>>(self, bound: T) -> Self

Limit the range to terms greater or equal to the bound

fn gt<T: AsRef<[u8]>>(self, bound: T) -> Self

Limit the range to terms strictly greater than the bound

fn lt<T: AsRef<[u8]>>(self, bound: T) -> Self

Limit the range to terms lesser or equal to the bound

fn le<T: AsRef<[u8]>>(self, bound: T) -> Self

Limit the range to terms lesser or equal to the bound

fn into_stream(self) -> Self::Streamer

Creates the stream corresponding to the range of terms defined using the TermStreamerBuilder.