Struct tantivy::tokenizer::NgramTokenizer

pub struct NgramTokenizer { /* fields omitted */ }

Tokenize the text by splitting words into n-grams of the given size(s)

With this tokenizer, the position field expresses the starting offset of the ngram rather than the token offset.

Example 1: hello would be tokenized as (min_gram: 2, max_gram: 3, prefix_only: false)

Term 	 he 	 hel 	 el 	 ell 	 ll 	 llo 	 lo

Position 	 0 	 0 	 1 	 1 	 2 	 2 	 3

Offsets 	 0,2 	 0,3 	 1,3 	 1,4 	 2,4 	 2,5 	 3,5

Example 2: hello would be tokenized as (min_gram: 2, max_gram: 5, prefix_only: true)

Term 	 he 	 hel 	 hell 	 hello

Position 	 0 	 0 	 0 	 0

Offsets 	 0,2 	 0,3 	 0,4 	 0,5

Example

extern crate tantivy;
use tantivy::tokenizer::*;
use tantivy::tokenizer::assert_token;

let tokenizer = NgramTokenizer::new(2, 3, false);
let mut stream = tokenizer.token_stream("hello");

assert_token(stream.next().unwrap(), 0, "he", 0, 2);
assert_token(stream.next().unwrap(), 0, "hel", 0, 3);
assert_token(stream.next().unwrap(), 1, "el", 1, 3);
assert_token(stream.next().unwrap(), 1, "ell", 1, 4);
assert_token(stream.next().unwrap(), 2, "ll", 2, 4);
assert_token(stream.next().unwrap(), 2, "llo", 2, 5);
assert_token(stream.next().unwrap(), 3, "lo", 3, 5);
assert!(stream.next().is_none());

Methods

impl NgramTokenizer

pub fn new(
min_gram: usize,
max_gram: usize,
prefix_only: bool
) -> NgramTokenizer

Configures a new Ngram tokenizer