Crate tantivy

tantivy

Tantivy is a search engine library. Think Lucene, but in Rust.

A good place for you to get started is to check out the example code ( literate programming / source code)

Modules

collector
Collector module

directory
WORM directory abstraction.

fastfield
Fast fields is a column oriented storage.

merge_policy
Defines tantivy's merging strategy

postings
Postings module (also called inverted index)

query
Query module

schema

store
Tantivy's store is a compressed, row-oriented storage.

termdict
The term dictionary is one of the key datastructure of tantivy. It associates sorted terms to their respective posting list.

Macros

doc
doc! is a shortcut that helps building Document objects.

Structs

DocAddress
DocAddress contains all the necessary information to identify a document given a Searcher object.

Document
Tantivy's Document is the object that can be indexed and then searched for.

Error
The Error type.

Index
Tantivy's Search Index

IndexWriter
IndexWriter is the user entry-point to add document to an index.

Searcher
Holds a list of SegmentReaders ready for search.

Segment
A segment is a piece of the index.

SegmentId
Tantivy SegmentId.

SegmentMeta
SegmentMeta contains simple meta information about a segment.

SegmentReader
Entry point to access all of the datastructures of the Segment

Term
Term represents the value that the token can take.

TimerTree
Timer tree

Enums

ErrorKind
The kind of an error.

SegmentComponent
Enum describing each component of a tantivy segment. Each component is stored in its own file, using the pattern segment_uuid.component_extension, except the delete component that takes an segment_uuid.delete_opstamp.component_extension

SegmentPostingsOption
Object describing the amount of information required when reading a postings.

Traits

Directory
Write-once read many (WORM) abstraction for where tantivy's data should be stored.

DocSet
Represents an iterable set of sorted doc ids.

Postings
Postings (also called inverted list)

ResultExt
Additional methods for Result, for easy interaction with this crate.

Functions

i64_to_u64
Maps a i64 to u64

u64_to_i64
Reverse the mapping given by i64_to_u64.

version
Expose the current version of tantivy, as well whether it was compiled with the simd compression.

Type Definitions

DocId
A u32 identifying a document within a segment. Documents have their DocId assigned incrementally, as they are added in the segment.

Result
Tantivy result.

Score
A f32 that represents the relevance of the document to the query

SegmentLocalId
A SegmentLocalId identifies a segment. It only makes sense for a given searcher.