Module tantivy::aggregation::bucket

Module for all bucket aggregations.

BucketAggregations create buckets of documents BucketAggregation.

Results of final buckets are BucketResult. Results of intermediate buckets are IntermediateBucketResult

Structs

CustomOrder

Set the order. target is either “_count”, “_key”, or the name of a metric sub_aggregation.

HistogramAggregation

Histogram is a bucket aggregation, where buckets are created dynamically for given interval. Each document value is rounded down to its bucket.

HistogramBounds

Used to set extended or hard bounds on the histogram.

RangeAggregation

Provide user-defined buckets to aggregate on. Two special buckets will automatically be created to cover the whole range of values. The provided buckets have to be continuous. During the aggregation, the values extracted from the fast_field field will be checked against each bucket range. Note that this aggregation includes the from value and excludes the to value for each range.

RangeAggregationRange

The range for one range bucket.

SegmentTermCollector

The collector puts values from the fast field into the correct buckets and does a conversion to the correct datatype.

TermsAggregation

Creates a bucket for every unique term and counts the number of occurences. Note that doc_count in the response buckets equals term count here.

Enums

Order

Order for buckets in a bucket aggregation.

OrderTarget

Order property by which to apply the order