Module tantivy::collector

Defines how the documents matching a search query should be processed.

Structs

CountCollector
CountCollector collector only counts how many documents match the query.

FacetCollector
Collector for faceting

MultiCollector
Multicollector makes it possible to collect on more than one collector. It should only be used for use cases where the Collector types is unknown at compile time. If the type of the collectors is known, you should prefer to use ChainedCollector.

TopCollector
The Top Collector keeps track of the K documents with the best scores.

Traits

Collector
Collectors are in charge of collecting and retaining relevant information from the document found and scored by the query.

Functions

chain
Creates a ChainedCollector