Struct tantivy::collector::MultiCollector

pub struct MultiCollector<'a> { /* fields omitted */ }

Multicollector makes it possible to collect on more than one collector. It should only be used for use cases where the Collector types is unknown at compile time. If the type of the collectors is known, you should prefer to use ChainedCollector.

Methods

impl<'a> MultiCollector<'a>

fn from(collectors: Vec<&'a mut Collector>) -> MultiCollector

Constructor