Struct tantivy::collector::TopCollector

pub struct TopCollector { /* fields omitted */ }

The Top Collector keeps track of the K documents with the best scores.

The implementation is based on a BinaryHeap. The theorical complexity is O(n log K).

Methods

impl TopCollector

fn with_limit(limit: usize) -> TopCollector

Creates a top collector, with a number of documents equal to "limit".

Panics

The method panics if limit is 0

fn docs(&self) -> Vec<DocAddress>

Returns K best documents sorted in decreasing order.

Calling this method triggers the sort. The result of the sort is not cached.

fn score_docs(&self) -> Vec<(Score, DocAddress)>

Returns K best ScoredDocument sorted in decreasing order.

Calling this method triggers the sort. The result of the sort is not cached.

fn at_capacity(&self) -> bool

Return true iff at least K documents have gone through the collector.