Trait tantivy::collector::Collector

pub trait Collector: Sync {
    type Fruit: Fruit;
    type Child: SegmentCollector<Fruit = Self::Fruit>;
    fn for_segment(
        &self, 
        segment_local_id: SegmentLocalId, 
        segment: &SegmentReader
    ) -> Result<Self::Child>;

    fn requires_scoring(&self) -> bool;

    fn merge_fruits(
        &self, 
        segment_fruits: Vec<Self::Fruit>
    ) -> Result<Self::Fruit>;
}

Collectors are in charge of collecting and retaining relevant information from the document found and scored by the query.

For instance,

keeping track of the top 10 best documents

computing a breakdown over a fast field

computing the number of documents matching the query

Our search index is in fact a collection of segments, so a Collector trait is actually more of a factory to instance SegmentCollectors for each segments.

The collection logic itself is in the SegmentCollector.

Segments are not guaranteed to be visited in any specific order.

Associated Types

type Fruit: Fruit

Fruit is the type for the result of our collection. e.g. usize for the Count collector.

type Child: SegmentCollector<Fruit = Self::Fruit>

Type of the SegmentCollector associated to this collector.

Loading content...
Required methods

fn for_segment(
&self,
segment_local_id: SegmentLocalId,
segment: &SegmentReader
) -> Result<Self::Child>

set_segment is called before beginning to enumerate on this segment.

fn requires_scoring(&self) -> bool

Returns true iff the collector requires to compute scores for documents.

fn merge_fruits(&self, segment_fruits: Vec<Self::Fruit>) -> Result<Self::Fruit>

Combines the fruit associated to the collection of each segments into one fruit.

Loading content...