Enum tantivy::Error

pub enum Error {
    PathDoesNotExist(PathBuf),
    FileAlreadyExists(PathBuf),
    IOError(Error),
    Poisoned,
    CorruptedFile(PathBuf, Box<Error + Send>),
    InvalidArgument(String),
    ErrorInThread(String),
}

Generic tantivy error.

Any specialized error return in tantivy can be converted in tantivy::Error.

Variants
PathDoesNotExist(PathBuf)

Path does not exist.

FileAlreadyExists(PathBuf)

File already exists, this is a problem when we try to write into a new file.

IOError(Error)

IO Error

Poisoned

A thread holding the locked panicked and poisoned the lock.

CorruptedFile(PathBuf, Box<Error + Send>)

The data within is corrupted.

For instance, it contains invalid JSON.

InvalidArgument(String)

Invalid argument was passed by the user.

ErrorInThread(String)

An Error happened in one of the thread