Enum tantivy::SegmentComponent

pub enum SegmentComponent {
    POSTINGS,
    POSITIONS,
    FASTFIELDS,
    FIELDNORMS,
    TERMS,
    STORE,
    DELETE,
}

Enum describing each component of a tantivy segment. Each component is stored in its own file, using the pattern segment_uuid.component_extension, except the delete component that takes an segment_uuid.delete_opstamp.component_extension

Variants
POSTINGS

Postings (or inverted list). Sorted lists of document ids, associated to terms

POSITIONS

Positions of terms in each document.

FASTFIELDS

Column-oriented random-access storage of fields.

FIELDNORMS

Stores the sum of the length (in terms) of each field for each document. Field norms are stored as a special u64 fast field.

TERMS

Dictionary associating Terms to TermInfos which is simply an address into the postings file and the positions file.

STORE

Row-oriented, LZ4-compressed storage of the documents. Accessing a document from the store is relatively slow, as it requires to decompress the entire block it belongs to.

DELETE

Bitset describing which document of the segment is deleted.

Methods

impl SegmentComponent

Important traits for Iter<'a, T>
impl<'a, T> Iterator for Iter<'a, T> type Item = &'a T;

pub fn iterator() -> Iter<'static, SegmentComponent>

Iterates through the components.