Module tantivy::fastfield

Column oriented field storage for tantivy.

It is the equivalent of Lucene's DocValues.

Fast fields is a column-oriented fashion storage of tantivy.

It is designed for the fast random access of some document fields given a document id.

FastField are useful when a field is required for all or most of the DocSet : for instance for scoring, grouping, filtering, or faceting.

Fields have to be declared as FAST in the schema. Currently only 64-bits integers (signed or unsigned) are supported.

They are stored in a bit-packed fashion so that their memory usage is directly linear with the amplitude of the values stored.

Read access performance is comparable to that of an array lookup.

Structs

DeleteBitSet
Set of deleted DocIds.

FacetReader
The facet reader makes it possible to access the list of facets associated to a given document in a specific segment.

FastFieldNotAvailableError
FastFieldNotAvailableError is returned when the user requested for a fast field reader, and the field was not defined in the schema as a fast field.

FastFieldReader
Trait for accessing a fastfield.

FastFieldSerializer
FastFieldSerializer is in charge of serializing fastfields on disk.

FastFieldsWriter
The fastfieldswriter regroup all of the fast field writers.

IntFastFieldWriter
Fast field writer for ints. The fast field writer just keeps the values in memory.

MultiValueIntFastFieldReader
Reader for a multivalued u64 fast field.

MultiValueIntFastFieldWriter

Traits

FastValue
Trait for types that are allowed for fast fields: (u64 or i64).

Functions

write_delete_bitset
Write a delete BitSet

Type Definitions

Result
Result when trying to access a fast field reader.