Module tantivy::fastfield

Fast fields is a column oriented storage.

It is the equivalent of Lucene's DocValues.

Fast fields is a column-oriented fashion storage of tantivy.

It is designed for the fast random access of some document fields given a document id.

FastField are useful when a field is required for all or most of the DocSet : for instance for scoring, grouping, filtering, or facetting.

Fields have to be declared as FAST in the schema. Currently only 64-bits integers (signed or unsigned) are supported.

They are stored in a bitpacked fashion so that their memory usage is directly linear with the amplitude of the values stored.

Read access performance is comparable to that of an array lookup.

Structs

DeleteBitSet
Set of deleted DocIds.

FastFieldNotAvailableError
FastFieldNotAvailableError is returned when the user requested for a fast field reader, and the field was not defined in the schema as a fast field.

FastFieldSerializer
FastFieldSerializer is in charge of serializing fastfields on disk.

FastFieldsReader
The FastFieldsReader is the datastructure containing all of the fast fields' data.

FastFieldsWriter
The fastfieldswriter regroup all of the fast field writers.

I64FastFieldReader
FastFieldReader for signed 64-bits integers.

IntFastFieldWriter
Fast field writer for ints. The fast field writer just keeps the values in memory.

U64FastFieldReader
FastFieldReader for unsigned 64-bits integers.

Traits

FastFieldReader
Trait for accessing a fastfield.

Functions

write_delete_bitset
Write a delete BitSet

Type Definitions

Result
Result when trying to access a fast field reader.