Module tantivy::fastfield

Column oriented field storage for tantivy.

It is the equivalent of Lucene’s DocValues.

A fast field is a column-oriented fashion storage for tantivy.

It is designed for the fast random access of some document fields given a document id.

Fast fields are useful when a field is required for all or most of the DocSet: for instance for scoring, grouping, aggregation, filtering, or faceting.

Fields have to be declared as FAST in the schema. Currently supported fields are: u64, i64, f64, bytes and text.

Fast fields are stored in with different codecs. The best codec is detected automatically, when serializing.

Read access performance is comparable to that of an array lookup.

Structs

AliveBitSet

Set of alive DocIds.

BytesFastFieldReader

Reader for byte array fast fields

BytesFastFieldWriter

Writer for byte array (as in, any number of bytes per document) fast fields

CompositeFastFieldSerializer

CompositeFastFieldSerializer is in charge of serializing fastfields on disk.

FacetReader

The facet reader makes it possible to access the list of facets associated with a given document in a specific segment.

FastFieldNotAvailableError

FastFieldNotAvailableError is returned when the user requested for a fast field reader, and the field was not defined in the schema as a fast field.

FastFieldReaders

Provides access to all of the BitpackedFastFieldReader.

FastFieldsWriter

The FastFieldsWriter groups all of the fast field writers.

IntFastFieldWriter

Fast field writer for ints. The fast field writer just keeps the values in memory.

MultiValueIndex

Index to resolve value range for given doc_id. Starts at 0.

MultiValueU128FastFieldWriter

Writer for multi-valued (as in, more than one value per document) int fast field.

MultiValuedFastFieldReader

Reader for a multivalued u64 fast field.

MultiValuedFastFieldWriter

Writer for multi-valued (as in, more than one value per document) int fast field.

MultiValuedU128FastFieldReader

Reader for a multivalued u128 fast field.

Enums

FastFieldType

The fast field type

Traits

Column

Column provides columnar access on a field.

FastValue

Trait for types that are allowed for fast fields: (u64, i64 and f64, bool, DateTime).

Functions

intersect_alive_bitsets

Intersects two AliveBitSets in a new one. The two bitsets need to have the same max_value.

write_alive_bitset

Write an alive BitSet

Type Definitions

Result

Result when trying to access a fast field reader.