Struct tantivy::fastfield::FastFieldSerializer

pub struct FastFieldSerializer { /* fields omitted */ }

FastFieldSerializer is in charge of serializing fastfields on disk.

Fast fields are encoded using bit-packing.

FastFieldWriters are in charge of pushing the data to the serializer. The serializer expects to receive the following calls.

new_u64_fast_field(...)

add_val(...)

add_val(...)

add_val(...)

...

close_field()

new_u64_fast_field(...)

add_val(...)

...

close_field()

close()

Methods

impl FastFieldSerializer

pub fn from_write(write: WritePtr) -> Result<FastFieldSerializer>

Constructor

pub fn new_u64_fast_field(
&mut self,
field: Field,
min_value: u64,
max_value: u64
) -> Result<FastSingleFieldSerializer<CountingWriter<WritePtr>>>

Start serializing a new u64 fast field

pub fn new_u64_fast_field_with_idx(
&mut self,
field: Field,
min_value: u64,
max_value: u64,
idx: usize
) -> Result<FastSingleFieldSerializer<CountingWriter<WritePtr>>>

Start serializing a new u64 fast field

pub fn close(self) -> Result<()>

Closes the serializer

After this call the data must be persistently save on disk.