Struct tantivy::fastfield::FastFieldsWriter

pub struct FastFieldsWriter { /* fields omitted */ }

The fastfieldswriter regroup all of the fast field writers.

Methods

impl FastFieldsWriter

pub fn from_schema(schema: &Schema) -> FastFieldsWriter

Create all FastFieldWriter required by the schema.

pub fn get_field_writer(
&mut self,
field: Field
) -> Option<&mut IntFastFieldWriter>

Get the FastFieldWriter associated to a field.

pub fn get_multivalue_writer(
&mut self,
field: Field
) -> Option<&mut MultiValueIntFastFieldWriter>

Returns the fast field multi-value writer for the given field.

Returns None if the field does not exist, or is not configured as a multivalued fastfield in the schema.

pub fn get_bytes_writer(
&mut self,
field: Field
) -> Option<&mut BytesFastFieldWriter>

Returns the bytes fast field writer for the given field.

Returns None if the field does not exist, or is not configured as a bytes fastfield in the schema.

pub fn add_document(&mut self, doc: &Document)

Indexes all of the fastfields of a new document.

pub fn serialize(
&self,
serializer: &mut FastFieldSerializer,
mapping: &HashMap<Field, HashMap<u64, TermOrdinal>>
) -> Result<()>

Serializes all of the FastFieldWriters by pushing them in order to the fast field serializer.