Struct tantivy::fastfield::FastFieldsWriter

pub struct FastFieldsWriter { /* fields omitted */ }

The fastfieldswriter regroup all of the fast field writers.

Methods

impl FastFieldsWriter

fn from_schema(schema: &Schema) -> FastFieldsWriter

Create all FastFieldWriter required by the schema.

fn new(fields: Vec<Field>) -> FastFieldsWriter

Returns a FastFieldsWriter with a IntFastFieldWriter for each of the field given in argument.

fn get_field_writer(&mut self, field: Field) -> Option<&mut IntFastFieldWriter>

Get the FastFieldWriter associated to a field.

fn add_document(&mut self, doc: &Document)

Indexes all of the fastfields of a new document.

fn serialize(&self, serializer: &mut FastFieldSerializer) -> Result<()>

Serializes all of the FastFieldWriters by pushing them in order to the fast field serializer.

fn fill_val_up_to(&mut self, doc: DocId)

Ensures all of the fast field writers have reached doc. (included)

The missing values will be filled with 0.