Struct tantivy::fastfield::U64FastFieldReader

pub struct U64FastFieldReader { /* fields omitted */ }

FastFieldReader for unsigned 64-bits integers.

Methods

impl U64FastFieldReader

pub fn min_value(&self) -> u64

Returns the minimum value for this fast field.

The min value does not take in account of possible deleted document, and should be considered as a lower bound of the actual minimum value.

pub fn max_value(&self) -> u64

Returns the maximum value for this fast field.

The max value does not take in account of possible deleted document, and should be considered as an upper bound of the actual maximum value.