Struct tantivy::postings::BlockSegmentPostings

pub struct BlockSegmentPostings { /* fields omitted */ }

BlockSegmentPostings is a cursor iterating over blocks of documents.

Warning

While it is useful for some very specific high-performance use cases, you should prefer using SegmentPostings for most usage.

Methods

impl BlockSegmentPostings

pub fn doc_freq(&self) -> usize

Returns the document frequency associated to this block postings.

This doc_freq is simply the sum of the length of all of the blocks length, and it does not take in account deleted documents.

Important traits for &'a mut [u8]
impl<'a> Write for &'a mut [u8]impl<'a> Read for &'a [u8]

pub fn docs(&self) -> &[DocId]

Returns the array of docs in the current block.

Before the first call to .advance(), the block returned by .docs() is empty.

pub fn doc(&self, idx: usize) -> u32

Return the document at index idx of the block.

Important traits for &'a mut [u8]
impl<'a> Write for &'a mut [u8]impl<'a> Read for &'a [u8]

pub fn freqs(&self) -> &[u32]

Return the array of term freq in the block.

pub fn freq(&self, idx: usize) -> u32

Return the frequency at index idx of the block.

pub fn advance(&mut self) -> bool

Advance to the next block.

Returns false iff there was no remaining blocks.

pub fn empty() -> BlockSegmentPostings

Returns an empty segment postings object