Struct tantivy::postings::SegmentPostings

pub struct SegmentPostings { /* fields omitted */ }

SegmentPostings represents the inverted list or postings associated to a term in a Segment.

As we iterate through the SegmentPostings, the frequencies are optionally decoded. Positions on the other hand, are optionally entirely decoded upfront.

Methods

impl SegmentPostings

pub fn create_from_docs(docs: &[u32]) -> SegmentPostings

Creates a segment postings object with the given documents and no frequency encoded.

This method is mostly useful for unit tests.

It serializes the doc ids using tantivy's codec and returns a SegmentPostings object that embeds a buffer with the serialized data.

pub fn from_block_postings(
segment_block_postings: BlockSegmentPostings,
delete_bitset: DeleteBitSet,
positions_stream_opt: Option<CompressedIntStream>
) -> SegmentPostings

Reads a Segment postings from an &[u8]

len - number of document in the posting lists.

data - data array. The complete data is not necessarily used.

freq_handler - the freq handler is in charge of decoding frequencies and/or positions

pub fn empty() -> SegmentPostings

Returns an empty segment postings object