Struct tantivy::postings::SegmentPostings

pub struct SegmentPostings<'a> { /* fields omitted */ }

SegmentPostings represents the inverted list or postings associated to a term in a Segment.

As we iterate through the SegmentPostings, the frequencies are optionally decoded. Positions on the other hand, are optionally entirely decoded upfront.

Methods

impl<'a> SegmentPostings<'a>

fn from_data(len: u32,
data: &'a [u8],
freq_handler: FreqHandler)
-> SegmentPostings<'a>

Reads a Segment postings from an &[u8]

len - number of document in the posting lists.

data - data array. The complete data is not necessarily used.

freq_handler - the freq handler is in charge of decoding frequencies and/or positions

fn empty() -> SegmentPostings<'static>

Returns an empty segment postings object