Module tantivy::query

Query module

Structs

DAATMultiTermScorer
Document-At-A-Time multi term scorer.

Explanation
Tree representing the expression of the score of document. The explanation is organized as follows.

MultiTermQuery
Query involving one or more terms.

QueryParser
Tantivy's Query parser

SimilarityExplainer
Wrapper over a similarity used to run explain

TfIdf
TfIdf is the default pertinence score in tantivy.

Enums

Occur
Defines whether a term in a query must be present, should be present or must not be present.

ParsingError
Possible error that may happen when parsing a query.

Traits

MultiTermAccumulator
Accumulator of the matching terms information

Query
Queries represent the query of the user, and are in charge of the logic defining the set of documents that should be sent to the collector, as well as the way to score the documents.

Scorer
Scored DocSet

Similarity
Similarity score