Module tantivy::query

Query module

Structs

BooleanQuery
The boolean query combines a set of queries

EmptyScorer
EmptyScorer is a dummy Scorer in which no document matches.

OccurFilter
An OccurFilter represents a filter over a bitset of

PhraseQuery
PhraseQuery matches a specific sequence of word. For instance the phrase query for "part time" will match the sentence

QueryParser
Tantivy's Query parser

TermQuery
A Term query matches all of the documents containing a specific term.

Enums

Occur
Defines whether a term in a query must be present, should be present or must not be present.

QueryParserError
Possible error that may happen when parsing a query.

Traits

Query
Query trait are in charge of defining :

Scorer
Scored set of documents matching a query within a specific segment.

Weight
A Weight is the specialization of a Query for a given set of segments.