Enum tantivy::query::QueryParserError

pub enum QueryParserError {
    SyntaxError,
    FieldDoesNotExist(String),
    ExpectedInt(ParseIntError),
    AllButQueryForbidden,
    NoDefaultFieldDeclared,
    FieldNotIndexed(String),
    FieldDoesNotHavePositionsIndexed(String),
    UnknownTokenizer(String, String),
    RangeMustNotHavePhrase,
    DateFormatError(ParseError),
}

Possible error that may happen when parsing a query.

Variants
SyntaxError

Error in the query syntax

FieldDoesNotExist(String)

FieldDoesNotExist(field_name: String) The query references a field that is not in the schema

ExpectedInt(ParseIntError)

The query contains a term for a u64-field, but the value is not a u64.

AllButQueryForbidden

It is forbidden queries that are only "excluding". (e.g. -title:pop)

NoDefaultFieldDeclared

If no default field is declared, running a query without any field specified is forbbidden.

FieldNotIndexed(String)

The field searched for is not declared as indexed in the schema.

FieldDoesNotHavePositionsIndexed(String)

A phrase query was requested for a field that does not have any positions indexed.

UnknownTokenizer(String, String)

The tokenizer for the given field is unknown The two argument strings are the name of the field, the name of the tokenizer

RangeMustNotHavePhrase

The query contains a range query with a phrase as one of the bounds. Only terms can be used as bounds.

DateFormatError(ParseError)

The format for the date field is not RFC 3339 compliant.