Enum tantivy::query::QueryParserError

pub enum QueryParserError {
    SyntaxError,
    FieldDoesNotExist(String),
    ExpectedU32(String, String),
    AllButQueryForbidden,
    NoDefaultFieldDeclared,
}

Possible error that may happen when parsing a query.

Variants
SyntaxError

Error in the query syntax

FieldDoesNotExist(String)

FieldDoesNotExist(field_name: String) The query references a field that is not in the schema

ExpectedU32(String, String)

ExpectedU32(field_name: String, field_value: String) The query contains a term for a u32-field, but the value is not a u32.

AllButQueryForbidden

It is forbidden queries that are only "excluding". (e.g. -title:pop)

NoDefaultFieldDeclared

If no default field is declared, running a query without any field specified is forbbidden.