Struct tantivy::query::BitSetDocSet

pub struct BitSetDocSet { /* fields omitted */ }

A BitSetDocSet makes it possible to iterate through a bitset as if it was a DocSet.

Implementation detail

Skipping is relatively fast here as we can directly point to the right tiny bitset bucket.

TODO: Consider implementing a BitTreeSet in order to advance faster when the bitset is sparse