Struct tantivy::query::BooleanQuery

pub struct BooleanQuery { /* fields omitted */ }

The boolean query combines a set of queries

The documents matched by the boolean query are those which

match all of the sub queries associated with the Must occurence

match none of the sub queries associated with the MustNot occurence.

match at least one of the subqueries that is not a MustNot occurence.

Methods

impl BooleanQuery

pub fn new_multiterms_query(terms: Vec<Term>) -> BooleanQuery

Helper method to create a boolean query matching a given list of terms. The resulting query is a disjunction of the terms.