Struct tantivy::query::MoreLikeThisQueryBuilder

pub struct MoreLikeThisQueryBuilder { /* private fields */ }

The builder for more-like-this query

Implementations

impl MoreLikeThisQueryBuilder

pub fn with_min_doc_frequency(self, value: u64) -> Self

Sets the minimum document frequency.

The resulting query will ignore words which do not occur in at least this many docs.

pub fn with_max_doc_frequency(self, value: u64) -> Self

Sets the maximum document frequency.

The resulting query will ignore words which occur in more than this many docs.

pub fn with_min_term_frequency(self, value: usize) -> Self

Sets the minimum term frequency.

The resulting query will ignore words less frequent that this number.

pub fn with_max_query_terms(self, value: usize) -> Self

Sets the maximum query terms.

The resulting query will not return a query with more clause than this.

pub fn with_min_word_length(self, value: usize) -> Self

Sets the minimum word length.

The resulting query will ignore words shorter than this length.

pub fn with_max_word_length(self, value: usize) -> Self

Sets the maximum word length.

The resulting query will ignore words longer than this length.

pub fn with_boost_factor(self, value: f32) -> Self

Sets the boost factor

The boost factor used by the resulting query for boosting terms.

pub fn with_stop_words(self, value: Vec<String>) -> Self

Sets the set of stop words

The resulting query will ignore these set of words.

pub fn with_document(self, doc_address: DocAddress) -> MoreLikeThisQuery

Sets the document address Returns the constructed MoreLikeThisQuery

This document will be used to collect field values, extract frequent terms needed for composing the query.

Note that field values will only be collected from stored fields in the index. You can construct your own field values from any source.

pub fn with_document_fields( self, doc_fields: Vec<(Field, Vec<Value>)> ) -> MoreLikeThisQuery

Sets the document fields Returns the constructed MoreLikeThisQuery

This represents the list field values possibly collected from multiple documents that will be used to compose the resulting query. This interface is meant to be used when you want to provide your own set of fields not necessarily from a specific document.