Struct tantivy::query::RangeQuery

pub struct RangeQuery { /* fields omitted */ }

RangeQuery match all documents that have at least one term within a defined range.

Matched document will all get a constant Score of one.

Implementation

The current implement will iterate over the terms within the range and append all of the document cross into a BitSet.

Example

let searcher = index.searcher();

let docs_in_the_sixties = RangeQuery::new_u64(year_field, 1960..1970);

// ... or `1960..=1969` if inclusive range is enabled.
let mut count_collector = CountCollector::default();
docs_in_the_sixties.search(&*searcher, &mut count_collector)?;

let num_60s_books = count_collector.count();

Methods

impl RangeQuery

pub fn new_i64<TRangeArgument: RangeArgument<i64>>(
field: Field,
range: TRangeArgument
) -> RangeQuery

Create a new RangeQuery over a i64 field.

pub fn new_u64<TRangeArgument: RangeArgument<u64>>(
field: Field,
range: TRangeArgument
) -> RangeQuery

Create a new RangeQuery over a u64 field.

pub fn new_str<'b, TRangeArgument: RangeArgument<&'b str>>(
field: Field,
range: TRangeArgument
) -> RangeQuery

Create a new RangeQuery over a Str field.