Trait tantivy::query::Query

pub trait Query {
    fn search<C: Collector>(&self,
                        searcher: &Searcher,
                        collector: &mut C)
                        -> Result<TimerTree>;
    fn explain(&self,
           searcher: &Searcher,
           doc_address: &DocAddress)
           -> Result<Explanation>;
}

Queries represent the query of the user, and are in charge of the logic defining the set of documents that should be sent to the collector, as well as the way to score the documents.

Required Methods

fn search<C: Collector>(&self,
searcher: &Searcher,
collector: &mut C)
-> Result<TimerTree>

Perform the search operation

fn explain(&self,
searcher: &Searcher,
doc_address: &DocAddress)
-> Result<Explanation>

Explain the score of a specific document