Trait tantivy::query::Scorer

pub trait Scorer: DocSet {
    fn score(&self) -> Score;

    fn collect(&mut self, collector: &mut Collector) { ... }
}

Scored set of documents matching a query within a specific segment.

See Query.

Required Methods

fn score(&self) -> Score

Returns the score.

This method will perform a bit of computation and is not cached.

Provided Methods

fn collect(&mut self, collector: &mut Collector)

Consumes the complete DocSet and push the scored documents to the collector.