Module tantivy::schema

Structs

Document
Document are really just a list of field values.

Field
Field handle

FieldOptions

FieldValue

Schema
Tantivy has a very strict schema. You need to specify in advance, whether a field is indexed or not, stored or not, and RAM-based or not.

Term

Constants

FAST
Fast field are used for field you need to access many times during collection. (e.g: for sort, aggregates).

INDEXED_TEXT
The field will be tokenized and indexed

STORED
A stored fields of a document can be retrieved given its DocId. Stored field are stored together and LZ4 compressed. Reading the stored fields of a document is relatively slow. (100 microsecs)

Type Definitions

DocId
u32 identifying a document within a segment. Document gets their doc id assigned incrementally, as they are added in the segment.