Enum tantivy::schema::FieldType

pub enum FieldType {
    Str(TextOptions),
    U64(IntOptions),
    I64(IntOptions),
    HierarchicalFacet,
}

A FieldType describes the type (text, u64) of a field as well as how it should be handled by tantivy.

Variants
Str(TextOptions)

String field type configuration

U64(IntOptions)

Unsigned 64-bits integers field type configuration

I64(IntOptions)

Signed 64-bits integers 64 field type configuration

HierarchicalFacet

Hierachical Facet

Methods

impl FieldType

pub fn value_type(&self) -> Type

Returns the value type associated for this field.

pub fn is_indexed(&self) -> bool

returns true iff the field is indexed.

pub fn get_index_record_option(&self) -> Option<IndexRecordOption>

Given a field configuration, return the maximal possible IndexRecordOption available.

If the field is not indexed, then returns None.

pub fn value_from_json(
&self,
json: &JsonValue
) -> Result<Value, ValueParsingError>

Parses a field value from json, given the target FieldType.

Tantivy will not try to cast values. For instance, If the json value is the integer 3 and the target field is a Str, this method will return an Error.