Enum tantivy::schema::FieldType

pub enum FieldType {
    Str(TextOptions),
    U64(IntOptions),
    I64(IntOptions),
}

A FieldType describes the type (text, u64) of a field as well as how it should be handled by tantivy.

Variants
Str(TextOptions)

String field type configuration

U64(IntOptions)

Unsigned 64-bits integers field type configuration

I64(IntOptions)

Signed 64-bits integers 64 field type configuration

Methods

impl FieldType

fn is_indexed(&self) -> bool

returns true iff the field is indexed.

fn get_segment_postings_option(&self) -> Option<SegmentPostingsOption>

Given a field configuration, return the maximal possible SegmentPostingsOption available.

If the field is not indexed, then returns None.

fn value_from_json(&self, json: &JsonValue) -> Result<Value, ValueParsingError>

Parses a field value from json, given the target FieldType.

Tantivy will not try to cast values. For instance, If the json value is the integer 3 and the target field is a Str, this method will return an Error.