Enum tantivy::schema::FieldType

pub enum FieldType {
    Str(TextOptions),
    U64(NumericOptions),
    I64(NumericOptions),
    F64(NumericOptions),
    Bool(NumericOptions),
    Date(DateOptions),
    Facet(FacetOptions),
    Bytes(BytesOptions),
    JsonObject(JsonObjectOptions),
    IpAddr(IpAddrOptions),
}

A FieldType describes the type (text, u64) of a field as well as how it should be handled by tantivy.

Variants

Str(TextOptions)

String field type configuration

U64(NumericOptions)

Unsigned 64-bits integers field type configuration

I64(NumericOptions)

Signed 64-bits integers 64 field type configuration

F64(NumericOptions)

64-bits float 64 field type configuration

Bool(NumericOptions)

Bool field type configuration

Date(DateOptions)

Signed 64-bits Date 64 field type configuration,

Facet(FacetOptions)

Hierarchical Facet

Bytes(BytesOptions)

Bytes (one per document)

JsonObject(JsonObjectOptions)

Json object

IpAddr(IpAddrOptions)

IpAddr field

Implementations

impl FieldType

pub fn value_type(&self) -> Type

Returns the value type associated for this field.

pub fn is_ip_addr(&self) -> bool

returns true if this is an ip address field

pub fn is_date(&self) -> bool

returns true if this is an date field

pub fn is_indexed(&self) -> bool

returns true if the field is indexed.

pub fn index_record_option(&self) -> Option<IndexRecordOption>

Returns the index record option for the field.

If the field is not indexed, returns None.

pub fn is_fast(&self) -> bool

returns true if the field is fast.

pub fn fastfield_cardinality(&self) -> Option<Cardinality>

returns true if the field is fast.

pub fn has_fieldnorms(&self) -> bool

returns true if the field is normed (see fieldnorms).

pub fn get_index_record_option(&self) -> Option<IndexRecordOption>

Given a field configuration, return the maximal possible IndexRecordOption available.

For the Json object, this does not necessarily mean it is the index record option level is available for all terms. (Non string terms have the Basic indexing option at most.)

If the field is not indexed, then returns None.

pub fn value_from_json( &self, json: JsonValue ) -> Result<Value, ValueParsingError>

Parses a field value from json, given the target FieldType.

Tantivy will not try to cast values. For instance, If the json value is the integer 3 and the target field is a Str, this method will return an Error.