Enum tantivy::schema::IndexRecordOption

pub enum IndexRecordOption {
    Basic,
    WithFreqs,
    WithFreqsAndPositions,
}

IndexRecordOption describes an amount information associated to a given indexed field.

It is both used to:

describe in the schema the amount of information that should be retained during indexing (See TextFieldIndexing.html.set_index_option)

to request for a given amount of information to be decoded as one goes through a posting list. (See InvertedIndexReader.read_postings)

Variants
Basic

records only the DocIds

WithFreqs

records the document ids as well as the term frequency. The term frequency can help giving better scoring of the documents.

WithFreqsAndPositions

records the document id, the term frequency and the positions of the occurences in the document. Positions are required to run PhraseQueries.

Methods

impl IndexRecordOption

pub fn is_termfreq_enabled(&self) -> bool

Returns true iff the term frequency will be encoded.

pub fn is_position_enabled(&self) -> bool

Returns true iff the term positions within the document are stored as well.

pub fn has_freq(&self) -> bool

Returns true iff this option includes encoding term frequencies.

pub fn has_positions(&self) -> bool

Returns true iff this option include encoding term positions.