Enum tantivy::schema::Value

pub enum Value {
    Str(String),
    U64(u64),
    I64(i64),
}

Value represents the value of a any field. It is an enum over all over all of the possible field type.

Variants
Str(String)

The str type is used for any text information.

U64(u64)

Unsigned 64-bits Integer u64

I64(i64)

Signed 64-bits Integer i64

Methods

impl Value

fn text(&self) -> &str

Returns the text value, provided the value is of the Str type.

Panics

If the value is not of type Str

fn u64_value(&self) -> u64

Returns the u64-value, provided the value is of the U64 type.

Panics

If the value is not of type U64

fn i64_value(&self) -> i64

Returns the i64-value, provided the value is of the I64 type.

Panics

If the value is not of type I64