Struct tantivy::schema::Facet

pub struct Facet(_);

A Facet represent a point in a given hierarchy.

They are typically represented similarly to a filepath. For instance, an e-commerce website could have a Facet for /electronics/tv_and_video/led_tv.

A document can be associated to any number of facets. The hierarchy implicitely imply that a document belonging to a facet also belongs to the ancestor of its facet. In the example above, /electronics/tv_and_video/ and /electronics.

Methods

impl Facet

pub fn root() -> Facet

Returns a new instance of the "root facet" Equivalent to /.

pub fn is_root(&self) -> bool

Returns true iff the facet is the root facet /.

Important traits for &'a mut [u8]
impl<'a> Write for &'a mut [u8]impl<'a> Read for &'a [u8]

pub fn encoded_bytes(&self) -> &[u8]

Returns a binary representation of the facet.

In this representation, 0u8 is used as a separator and the string parts of the facet are unescaped. (The first / is not encoded at all).

This representation has the benefit of making it possible to express "being a child of a given facet" as a range over the term ordinals.

pub fn from_text<T: ?Sized>(path: &T) -> Facet where
T: AsRef<str>,

Parse a text representation of a facet.

It is conceptually, if one of the steps of this path contains a / or a \, it should be escaped using an anti-slash /.

pub fn from_path<Path>(path: Path) -> Facet where
Path: IntoIterator,
Path::Item: ToString,

Returns a Facet from an iterator over the different steps of the facet path.

The steps are expected to be unescaped.

pub fn is_prefix_of(&self, other: &Facet) -> bool

Returns true iff other is a subfacet of self.