Struct tantivy::schema::FieldEntry

pub struct FieldEntry { /* fields omitted */ }

A FieldEntry represents a field and its configuration. Schema are a collection of FieldEntry

It consists of - a field name - a field type, itself wrapping up options describing how the field should be indexed.

Methods

impl FieldEntry

fn new_text(field_name: String, field_type: TextOptions) -> FieldEntry

Creates a new u32 field entry in the schema, given a name, and some options.

fn new_u32(field_name: String, field_type: U32Options) -> FieldEntry

Creates a new u32 field entry in the schema, given a name, and some options.

fn name(&self) -> &String

Returns the name of the field

fn field_type(&self) -> &FieldType

Returns the field type

fn is_indexed(&self) -> bool

Returns true iff the field is indexed

fn is_u32_fast(&self) -> bool

Returns true iff the field is a u32 fast field

fn is_stored(&self) -> bool

Returns true iff the field is stored