Struct tantivy::schema::Schema

pub struct Schema {
    // some fields omitted
}

Tantivy has a very strict schema. You need to specify in advance, whether a field is indexed or not, stored or not, and RAM-based or not.

This is done by creating a schema object, and setting up the fields one by one. It is for the moment impossible to remove fields.

Examples

use tantivy::Schema;
...
fn create_schema() -> Schema {
  let mut schema = Schema::new();
  let str_fieldtype = FieldOptions::new();
  let text_fieldtype = FieldOptions::new().set_tokenized_indexed();
  let id_field = schema.add_field("id", &str_fieldtype);
  let url_field = schema.add_field("url", &str_fieldtype);
  let body_field = schema.add_field("body", &text_fieldtype);
  let id_field = schema.add_field("id", &str_fieldtype);
  let url_field = schema.add_field("url", &str_fieldtype);
  let title_field = schema.add_field("title", &text_fieldtype);
  let body_field = schema.add_field("body", &text_fieldtype);
  schema
}

Methods

impl Schema

fn new() -> Schema

Creates a new, empty schema.

fn get(&self, field_name: &str) -> Option<(Field, FieldOptions)>

Given a name, returns the field handle, as well as its associated FieldOptions

fn field(&self, fieldname: &str) -> Field

Returns the field options associated with a given name.

Panics

Panics if the field name does not exist. It is meant as an helper for user who created and control the content of their schema.

If panicking is not an option for you, you may use get(&self, field_name: &str).

fn field_options(&self, field: &Field) -> FieldOptions

Returns the field options associated to a field handle.

fn add_field(&mut self, field_name_str: &str, field_options: &FieldOptions) -> Field

Creates a new field. Return the associated field handle.

impl Schema

fn new() -> Schema

Creates a new, empty schema.

fn get(&self, field_name: &str) -> Option<(Field, FieldOptions)>

Given a name, returns the field handle, as well as its associated FieldOptions

fn field(&self, fieldname: &str) -> Field

Returns the field options associated with a given name.

Panics

Panics if the field name does not exist. It is meant as an helper for user who created and control the content of their schema.

If panicking is not an option for you, you may use get(&self, field_name: &str).

fn field_options(&self, field: &Field) -> FieldOptions

Returns the field options associated to a field handle.

fn add_field(&mut self, field_name_str: &str, field_options: &FieldOptions) -> Field

Creates a new field. Return the associated field handle.