Struct tantivy::schema::Schema

pub struct Schema(_);

Tantivy has a very strict schema. You need to specify in advance, whether a field is indexed or not, stored or not, and RAM-based or not.

This is done by creating a schema object, and setting up the fields one by one. It is for the moment impossible to remove fields.

Examples

use tantivy::schema::*;

let mut schema_builder = SchemaBuilder::default();
let id_field = schema_builder.add_text_field("id", STRING);
let title_field = schema_builder.add_text_field("title", TEXT);
let body_field = schema_builder.add_text_field("body", TEXT);
let schema = schema_builder.build();

Methods

impl Schema

pub fn get_field_entry(&self, field: Field) -> &FieldEntry

Return the FieldEntry associated to a Field.

pub fn get_field_name(&self, field: Field) -> &str

Return the field name for a given Field.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn fields(&self) -> &[FieldEntry]

Return the list of all the Fields.

pub fn get_field(&self, field_name: &str) -> Option<Field>

Returns the field options associated with a given name.

Panics

Panics if the field name does not exist. It is meant as an helper for user who created and control the content of their schema.

If panicking is not an option for you, you may use get(&self, field_name: &str).

pub fn to_named_doc(&self, doc: &Document) -> NamedFieldDocument

Create a named document off the doc.

pub fn to_json(&self, doc: &Document) -> String

Encode the schema in JSON.

Encoding a document cannot fail.

pub fn parse_document(
&self,
doc_json: &str
) -> Result<Document, DocParsingError>

Build a document object from a json-object.