Module tantivy::space_usage

Representations for the space usage of various parts of a Tantivy index.

This can be used programmatically, and will also be exposed in a human readable fashion in tantivy-cli.

One important caveat for all of this functionality is that none of it currently takes storage-level details into consideration. For example, if your file system block size is 4096 bytes, we can under-count actual resultant space usage by up to 4095 bytes per file.

Structs

FieldUsage
Represents space usage of a given field, breaking it down into the (field, index) pairs that comprise it.

PerFieldSpaceUsage
Represents space usage for all of the (field, index) pairs that appear in a CompositeFile.

SearcherSpaceUsage
Represents combined space usage of an entire searcher and its component segments.

SegmentSpaceUsage
Represents combined space usage for all of the large components comprising a segment.

StoreSpaceUsage
Represents space usage for the Store for this segment.

Enums

ComponentSpaceUsage
Enum containing any of the possible space usage results for segment components.

Type Definitions

ByteCount
Indicates space usage in bytes