Struct tantivy::Index

pub struct Index { /* private fields */ }

Search Index

Implementations

impl Index

pub fn builder() -> IndexBuilder

Creates a new builder.

pub fn exists(dir: &dyn Directory) -> Result<bool, OpenReadError>

Examines the directory to see if it contains an index.

Effectively, it only checks for the presence of the meta.json file.

pub fn search_executor(&self) -> &Executor

Accessor to the search executor.

This pool is used by default when calling searcher.search(...) to perform search on the individual segments.

By default the executor is single thread, and simply runs in the calling thread.

pub fn set_multithread_executor(&mut self, num_threads: usize) -> Result<()>

Replace the default single thread search executor pool by a thread pool with a given number of threads.

pub fn set_default_multithread_executor(&mut self) -> Result<()>

Replace the default single thread search executor pool by a thread pool with as many threads as there are CPUs on the system.

pub fn create_in_ram(schema: Schema) -> Index

Creates a new index using the RamDirectory.

The index will be allocated in anonymous memory. This is useful for indexing small set of documents for instances like unit test or temporary in memory index.

pub fn create_in_dir<P: AsRef<Path>>( directory_path: P, schema: Schema ) -> Result<Index>

Creates a new index in a given filepath. The index will use the MmapDirectory.

If a previous index was in this directory, then it returns a TantivyError::IndexAlreadyExists error.

pub fn open_or_create<T: Into<Box<dyn Directory>>>( dir: T, schema: Schema ) -> Result<Index>

Opens or creates a new index in the provided directory

pub fn create_from_tempdir(schema: Schema) -> Result<Index>

Creates a new index in a temp directory.

The index will use the MmapDirectory in a newly created directory. The temp directory will be destroyed automatically when the Index object is destroyed.

The temp directory is only used for testing the MmapDirectory. For other unit tests, prefer the RamDirectory, see: IndexBuilder::create_in_ram().

pub fn create<T: Into<Box<dyn Directory>>>( dir: T, schema: Schema, settings: IndexSettings ) -> Result<Index>

Creates a new index given an implementation of the trait Directory.

If a directory previously existed, it will be erased.

pub fn set_tokenizers(&mut self, tokenizers: TokenizerManager)

Setter for the tokenizer manager.

pub fn tokenizers(&self) -> &TokenizerManager

Accessor for the tokenizer manager.

pub fn tokenizer_for_field(&self, field: Field) -> Result<TextAnalyzer>

Get the tokenizer associated with a specific field.

pub fn reader(&self) -> Result<IndexReader>

Create a default IndexReader for the given index.

See [Index.reader_builder()].

pub fn reader_builder(&self) -> IndexReaderBuilder

Create a IndexReader for the given index.

Most project should create at most one reader for a given index. This method is typically called only once per Index instance.

pub fn open_in_dir<P: AsRef<Path>>(directory_path: P) -> Result<Index>

Opens a new directory from an index path.

pub fn new_segment_meta( &self, segment_id: SegmentId, max_doc: u32 ) -> SegmentMeta

Creates a new segment_meta (Advanced user only).

As long as the SegmentMeta lives, the files associated with the SegmentMeta are guaranteed to not be garbage collected, regardless of whether the segment is recorded as part of the index or not.

pub fn open<T: Into<Box<dyn Directory>>>(directory: T) -> Result<Index>

Open the index using the provided directory

pub fn load_metas(&self) -> Result<IndexMeta>

Reads the index meta file from the directory.

pub fn writer_with_num_threads( &self, num_threads: usize, overall_memory_arena_in_bytes: usize ) -> Result<IndexWriter>

Open a new index writer. Attempts to acquire a lockfile.

The lockfile should be deleted on drop, but it is possible that due to a panic or other error, a stale lockfile will be left in the index directory. If you are sure that no other IndexWriter on the system is accessing the index directory, it is safe to manually delete the lockfile.

num_threads defines the number of indexing workers that should work at the same time.

overall_memory_arena_in_bytes sets the amount of memory allocated for all indexing thread. Each thread will receive a budget of overall_memory_arena_in_bytes / num_threads.

Errors

If the lockfile already exists, returns Error::DirectoryLockBusy or an Error::IoError. If the memory arena per thread is too small or too big, returns TantivyError::InvalidArgument

pub fn writer(&self, memory_arena_num_bytes: usize) -> Result<IndexWriter>

Creates a multithreaded writer

Tantivy will automatically define the number of threads to use, but no more than 8 threads. overall_memory_arena_in_bytes is the total target memory usage that will be split between a given number of threads.

Errors

If the lockfile already exists, returns Error::FileAlreadyExists. If the memory arena per thread is too small or too big, returns TantivyError::InvalidArgument

pub fn settings(&self) -> &IndexSettings

Accessor to the index settings

pub fn settings_mut(&mut self) -> &mut IndexSettings

Accessor to the index settings

pub fn schema(&self) -> Schema

Accessor to the index schema

The schema is actually cloned.

pub fn searchable_segments(&self) -> Result<Vec<Segment>>

Returns the list of segments that are searchable

pub fn new_segment(&self) -> Segment

Creates a new segment.

pub fn directory(&self) -> &ManagedDirectory

Return a reference to the index directory.

pub fn directory_mut(&mut self) -> &mut ManagedDirectory

Return a mutable reference to the index directory.

pub fn searchable_segment_metas(&self) -> Result<Vec<SegmentMeta>>

Reads the meta.json and returns the list of SegmentMeta from the last commit.

pub fn searchable_segment_ids(&self) -> Result<Vec<SegmentId>>

Returns the list of segment ids that are searchable.

pub fn validate_checksum(&self) -> Result<HashSet<PathBuf>>

Returns the set of corrupted files