Struct tantivy::Searcher

pub struct Searcher { /* fields omitted */ }

Holds a list of SegmentReaders ready for search.

It guarantees that the Segment will not be removed before the destruction of the Searcher.

Methods

impl Searcher

pub fn index(&self) -> &Index

Returns the Index associated to the Searcher

pub fn doc(&self, doc_address: DocAddress) -> Result<Document>

Fetches a document from tantivy's store given a DocAddress.

The searcher uses the segment ordinal to route the the request to the right Segment.

pub fn schema(&self) -> &Schema

Access the schema associated to the index of this searcher.

pub fn num_docs(&self) -> u64

Returns the overall number of documents in the index.

pub fn doc_freq(&self, term: &Term) -> u64

Return the overall number of documents containing the given term.

pub fn segment_readers(&self) -> &[SegmentReader]

Return the list of segment readers

pub fn segment_reader(&self, segment_ord: u32) -> &SegmentReader

Returns the segment_reader associated with the given segment_ordinal

pub fn search<C: Collector>(
&self,
query: &dyn Query,
collector: &C
) -> Result<C::Fruit>

Runs a query on the segment readers wrapped by the searcher.

Search works as follows :

First the weight object associated to the query is created.

Then, the query loops over the segments and for each segment :

setup the collector and informs it that the segment being processed has changed.

creates a SegmentCollector for collecting documents associated to the segment

creates a Scorer object associated for this segment

iterate through the matched documents and push them to the segment collector.

Finally, the Collector merges each of the child collectors into itself for result usability by the caller.

pub fn search_with_executor<C: Collector>(
&self,
query: &dyn Query,
collector: &C,
executor: &Executor
) -> Result<C::Fruit>

Same as search(...) but multithreaded.

The current implementation is rather naive : multithreading is by splitting search into as many task as there are segments.

It is powerless at making search faster if your index consists in one large segment.

Also, keep in my multithreading a single query on several threads will not improve your throughput. It can actually hurt it. It will however, decrease the average response time.

pub fn field(&self, field: Field) -> FieldSearcher

Return the field searcher associated to a Field.

pub fn space_usage(&self) -> SearcherSpaceUsage

Summarize total space usage of this searcher.