Struct tantivy::Searcher

pub struct Searcher { /* fields omitted */ }

Holds a list of SegmentReaders ready for search.

It guarantees that the Segment will not be removed before the destruction of the Searcher.

Methods

impl Searcher

pub fn doc(&self, doc_address: &DocAddress) -> Result<Document>

Fetches a document from tantivy's store given a DocAddress.

The searcher uses the segment ordinal to route the the request to the right Segment.

pub fn num_docs(&self) -> DocId

Returns the overall number of documents in the index.

pub fn doc_freq(&self, term: &Term) -> u32

Return the overall number of documents containing the given term.

Important traits for &'a mut [u8]
impl<'a> Write for &'a mut [u8]impl<'a> Read for &'a [u8]

pub fn segment_readers(&self) -> &[SegmentReader]

Return the list of segment readers

pub fn segment_reader(&self, segment_ord: u32) -> &SegmentReader

Returns the segment_reader associated with the given segment_ordinal

pub fn search<C: Collector>(
&self,
query: &Query,
collector: &mut C
) -> Result<TimerTree>

Runs a query on the segment readers wrapped by the searcher

pub fn field(&self, field: Field) -> FieldSearcher

Return the field searcher associated to a Field.