Struct tantivy::SegmentReader

pub struct SegmentReader {
    // some fields omitted
}

Methods

impl SegmentReader

fn max_doc(&self) -> DocId

Returns the highest document id ever attributed in this segment + 1. Today, tantivy does not handle deletes so, it happens to also be the number of documents in the index.

fn open(segment: Segment) -> Result<SegmentReader>

Open a new segment for reading.

fn doc(&self, doc_id: &DocId) -> Result<Document>

Returns the document (or to be accurate, its stored field) bearing the given doc id. This method is slow and should seldom be called from within a collector.

fn search(&self, terms: &Vec<Term>) -> IntersectionPostings<SegmentPostings>

Returns the list of doc ids containing all of the given terms.