Struct tantivy::SegmentReader

pub struct SegmentReader { /* fields omitted */ }

Entry point to access all of the datastructures of the Segment

term dictionary

postings

store

fast field readers

field norm reader

The segment reader has a very low memory footprint, as close to all of the memory data is mmapped.

TODO fix not decoding docfreq

Methods

impl SegmentReader

pub fn max_doc(&self) -> DocId

Returns the highest document id ever attributed in this segment + 1. Today, tantivy does not handle deletes, so it happens to also be the number of documents in the index.

pub fn num_docs(&self) -> DocId

Returns the number of documents. Deleted documents are not counted.

Today, tantivy does not handle deletes so max doc and num_docs are the same.

pub fn schema(&self) -> &Schema

Returns the schema of the index this segment belongs to.

pub fn num_deleted_docs(&self) -> DocId

Return the number of documents that have been deleted in the segment.

pub fn has_deletes(&self) -> bool

Returns true iff some of the documents of the segment have been deleted.

pub fn fast_fields(&self) -> &FastFieldReaders

Accessor to a segment's fast field reader given a field.

Returns the u64 fast value reader if the field is a u64 field indexed as "fast".

Return a FastFieldNotAvailableError if the field is not declared as a fast field in the schema.

Panics

May panic if the index is corrupted.

pub fn facet_reader(&self, field: Field) -> Option<FacetReader>

Accessor to the FacetReader associated to a given Field.

pub fn get_fieldnorms_reader(&self, field: Field) -> FieldNormReader

Accessor to the segment's Field norms's reader.

Field norms are the length (in tokens) of the fields. It is used in the computation of the TfIdf.

They are simply stored as a fast field, serialized in the .fieldnorm file of the segment.

pub fn get_store_reader(&self) -> StoreReader

Accessor to the segment's StoreReader.

pub fn open(segment: &Segment) -> Result<SegmentReader>

Open a new segment for reading.

pub fn inverted_index(&self, field: Field) -> Arc<InvertedIndexReader>

Returns a field reader associated to the field given in argument. If the field was not present in the index during indexing time, the InvertedIndexReader is empty.

The field reader is in charge of iterating through the term dictionary associated to a specific field, and opening the posting list associated to any term.

pub fn segment_id(&self) -> SegmentId

Returns the segment id

pub fn delete_bitset(&self) -> Option<&DeleteBitSet>

Returns the bitset representing the documents that have been deleted.

pub fn is_deleted(&self, doc: DocId) -> bool

Returns true iff the doc is marked as deleted.

pub fn doc_ids_alive(&self) -> SegmentReaderAliveDocsIterator

Returns an iterator that will iterate over the alive document ids

pub fn space_usage(&self) -> SegmentSpaceUsage

Summarize total space usage of this segment.