Struct tantivy::termdict::TermMerger

pub struct TermMerger<'a> { /* fields omitted */ }

Given a list of sorted term streams, returns an iterator over sorted unique terms.

The item yield is actually a pair with

the term

a slice with the ordinal of the segments containing the terms.

Methods

impl<'a> TermMerger<'a>

pub fn new(streams: Vec<TermStreamerImpl<'a>>) -> TermMerger<'a>

Stream of merged term dictionary

pub fn advance(&mut self) -> bool

Advance the term iterator to the next term. Returns true if there is indeed another term False if there is none.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn key(&self) -> &[u8]

Returns the current term.

This method may be called iff advance() has been called before and "true" was returned.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn current_kvs(&self) -> &[HeapItem<'a>]

Returns the sorted list of segment ordinals that include the current term.

This method may be called iff advance() has been called before and "true" was returned.

pub fn next(&mut self) -> Option<Term<&[u8]>>

Iterates through terms